% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availability_message}
\alias{availability_message}
\title{Availability Message}
\usage{
availability_message(what, addtl = NULL)
}
\arguments{
\item{what}{(\code{character})\cr
Which clinical table is used.}

\item{addtl}{(\code{character})\cr
An additional, specific message to append to the generic message.}
}
\value{
Prints a message to the console.
}
\description{
Message to display to inform the user that the function being used does not
check clinical data availability.
}
