% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_filter.R
\name{cci_filter}
\alias{cci_filter}
\title{Filtering and Identify CCI intervention codes}
\usage{
cci_filter(dbcon)
}
\arguments{
\item{user_table}{(\code{data.frame} | \code{data.table})
Table containing all interventions of interest, where one corresponds to
CCI codes of patients. Please make sure the table contains a column named
\code{intervention_code} (\code{character}).}
}
\value{
\code{data.table}
This function returns a table containing the CCI codes of interest, together
with their corresponding definitions.
For each row in the output table, the following variables are returned:
\itemize{
\item \code{intervention_code}: CCI code that matches the filtration criteria
\item \code{CCI_description}: description of the CCI code
}
}
\description{
The \href{https://www.cihi.ca/en/cci-coding-structure}{Canadian Classification of Health Interventions (CCI) codes}
provide detailed documentation of all in-patient interventions in Canada.
CCI codes are alphanumeric codes with 5-10 characters, where a given character
defines the following features:
\enumerate{
\item Section (1 character): Broad realm of interventions (e.g., 1 = physical/physiological therapeutic interventions)
\item Group (2 characters): Subgroups within each Section, e.g., based on anatomy site (e.g., SC = spinal vertebrae)
\item Intervention (2 characters): Procedure/intervention types within a given Section (e.g., 75 = fusion)
\item Qualifier 1 (2 characters): Optional details, such as approach/technique used (e.g., LL = using open anterior approach)
\item Qualifier 2 (2 characters): Optional details, such as devices used/implanted (e.g., KD = fixation device)
\item Qualifier 3 (1 character): Optional details, such as tissues used (e.g., A = autograft)
}

The codes are structured hierarchically and the specific groups, interventions, and qualifiers vary across Sections.

The \code{cci_filter()} function provides a step-by-step search allowing users to extract a
subset of CCI codes matching the Section, Group(s), Intervention(s), and/or Qualifier(s)
that are of interest for a particular study.

At each step, the user is shown available categories and their corresponding descriptions,
which can be searched and sorted in the viewer pane. The user is then prompted to provide
the entries they want to include in the terminal. In some cases, broader drop-down menus are
shown allowing users to select multiple categories at once (e.g., all interventions targeting
the nervous system).
}
\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
  dbname = "db",
  host = "domain_name.ca",
  port = 1234,
  user = getPass("Enter user:"),
  password = getPass("password")
)
}
a <- cci_filter(cci_lookup)

}
