% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_group.R
\name{cci_group}
\alias{cci_group}
\title{Grouping CCI intervention codes}
\usage{
cci_group(cci_codes)
}
\arguments{
\item{cci_codes}{(\code{data.frame} | \code{data.table})
Table containing \code{intervention_code} column that lists all CCI intervention
codes of interest (e.g., from the \code{ipintervention} or \code{erintervention} table).}
}
\value{
\code{data.table}
Data table containing the CCI codes of interest, together with their
corresponding groupings as follows:
\itemize{
\item \code{section} (\code{numeric}): CCI section number (based on first character)
\item \code{section_descr} (\code{character}): Section description
\item \code{subsection} (\code{character}): CCI subsection (first 3 characters)
\item \code{subsection_descr} (\code{character}): Subsection description
}
}
\description{
The \href{https://www.cihi.ca/sites/default/files/document/guide-for-users-of-cci-manual-en.pdf}{Canadian Classification of Health Interventions (CCI) codes} provide a
detailed classification of all inpatient interventions in Canada,
with more than 17,000 unique, alphanumeric codes.

The \code{cci_group()} function categorizes CCI codes into broader, clinically
meaningful intervention categories based on each code's first 3 characters.

Broadly, intervention codes can be grouped into the following CCI Sections
based on the first digit:
\itemize{
\item 1: Physical/Physiological Therapeutic Interventions (1AA - 1ZZ)
\item 2: Diagnostic Interventions (2AA - 2ZZ)
\item 3: Diagnostic Imaging Interventions (3AF - 3ZZ)
\item 5: Obstetrical and Fetal Interventions (5AB - 5PD)
\item 6: Cognitive, Psychosocial and Sensory Therapeutic Interventions (6AA - 6VA)
\item 7: Other Healthcare Interventions (7SC - 7SJ)
\item 8: Therapeutic Interventions Strengthening the Immune System and/or Genetic Composition (8AA - 8ZZ)
}

Each section can further be broken down into more detailed subsections, based
on the CCI group (2nd-3rd character).For example, CCI codes in Sections 1-3
can be differentiated by anatomy region, such as:
\itemize{
\item Therapeutic Interventions on the Nervous System (1AA - 1BZ)
\item Therapeutic Interventions on the Eye and Ocular Adnexa (1CC - 1CZ)
\item Therapeutic Interventions on the Ear and Mastoid (1DA - 1DZ)
\item etc.
}

For a complete list of all subsections, see
\href{https://www.cihi.ca/en/overview-of-cci-sections-and-code-ranges}{CIHI CCI sections and code ranges}.
}
\examples{
cci_codes <- data.frame(intervention_code = c(
  "1MA52HA", "2PB70DA", "3SC10KM", "5LB08ZZ", "6KA02ME", "8AA70BABA"
))
res <- cci_group(cci_codes)

}
\references{
https://www.cihi.ca/en/overview-of-cci-sections-and-code-ranges
}
