% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_to_datatable}
\alias{coerce_to_datatable}
\title{Coerce to \code{data.table}}
\usage{
coerce_to_datatable(data)
}
\arguments{
\item{data(`data.frame`}{or \code{data.table})\cr
The data to check class of and coerce to \code{data.table} if necessary.}
}
\value{
(\code{data.table})\cr
The original data provided as an argument, but as a \code{data.table}.
}
\description{
Some \code{Rgemini} functions rely on \code{data.table} operations and assume the input is provided
in \code{data.table} format. If it  is not, coerce with message to ensure the function logic
works without breaking.
}
\examples{
mtcars <- coerce_to_datatable(mtcars)

}
