% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_index.R
\name{comorbidity_index}
\alias{comorbidity_index}
\title{Compute Comorbidity Index}
\usage{
comorbidity_index(
  ipdiag,
  erdiag,
  map,
  weights,
  at_admission = TRUE,
  raw_comorbidities = FALSE
)
}
\arguments{
\item{ipdiag}{(\code{data.table} or \code{data.frame})\cr
\code{ipdiagnosis} table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.}

\item{erdiag}{(\code{data.table} or \code{data.frame})\cr
\code{erdiagnosis} table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.
Typically, ER diagnoses should be included when deriving comorbidity in order
to increase sensitivity. However, in certain scenarios, users may choose to
only include IP diagnoses by specifying \code{erdiag = NULL}. This may be useful
when comparing cohorts with different rates of ER admissions.}

\item{map}{String denoting the mapping algorithm to be used (values are case-insensitive).
Possible values are the Charlson score with either ICD-10 or ICD-9-CM codes (\code{charlson_icd10_quan}, \code{charlson_icd9_quan}) and the Elixhauser score, again using either ICD-10 or ICD-9-CM (\code{elixhauser_icd10_quan}, \code{elixhauser_icd9_quan}).
These mapping are based on the paper by Quan et al. (2011).
It is also possible to obtain a Swedish (\code{charlson_icd10_se}) or Australian (\code{charlson_icd10_am}) modification of the Charlson score using ICD-10 codes.}

\item{weights}{A string denoting the weighting system to be used, which will depend on the mapping algorithm.

Possible values for the Charlson index are:
\itemize{
\item \code{charlson}, for the original weights by Charlson et al. (1987);
\item \code{quan}, for the revised weights by Quan et al. (2011).
}

Possible values for the Elixhauser score are:
\itemize{
\item \code{vw}, for the weights by van Walraven et al. (2009);
\item \code{swiss}, for the Swiss Elixhauser weights by Sharma et al. (2021).
}

Defaults to \code{NULL}, in which case an unweighted score will be used.}

\item{at_admission}{(\code{logical})\cr
Whether to calculate the score for all comorbidities or for only pre-admit comorbidities.}

\item{raw_comorbidities}{(\code{logical})\cr
Whether to output a \code{data.table} of raw comorbidities as opposed to pre-computed scores.}
}
\value{
(\code{data.table})\cr
By default, for each encounter, outputs the comorbidity score. If \code{raw_comorbidities} is \code{TRUE},
outputs a wide \code{data.table} with a column for each comorbidity for each encounter.
}
\description{
This is a generic function that is wrapped by \code{\link[=charlson_comorbidity_index]{charlson_comorbidity_index()}}
and \code{\link[=elixhauser_comorbidity_index]{elixhauser_comorbidity_index()}} and is not meant to be used directly.
The function is an interface to \code{\link[comorbidity:comorbidity]{comorbidity::comorbidity()}} for GEMINI data,
providing options for outputting raw comorbidities or comorbidities at admission.
}
\details{
A hierarchy of comorbidities is used when calculating the cormorbidity score, but not when outputting \code{raw_comorbidites}.
This affects comorbidities present in a patient with different degrees of severity. See documentation for
\code{\link[comorbidity:comorbidity]{comorbidity::comorbidity()}} for details.
}
