% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_sets}
\alias{compare_sets}
\title{Compare two sets to get the number of unique and common elements in each set}
\usage{
compare_sets(x, y, dates = FALSE)
}
\arguments{
\item{x}{('vector') \cr
The first set to be compared}

\item{y}{('vector') \cr
The second set to be compared}

\item{dates}{(\code{logical})\cr
If set to TRUE, will attempt to convert x & y into dates, and then compare. Set to FALSE by default.}
}
\value{
(\code{data.table})\cr
A table showing the number of elements in both vectors, in the first vector only, and in the second vector only
}
\description{
This function takes in two vectors and returns the number
of overlapping elements, along with the number of unique elements.
The function can compare sets of characters, numerics, and dates.
}
\examples{
compare_sets(c(1:10), c(5:10), dates = FALSE)
}
