% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_datetime}
\alias{convert_datetime}
\title{Convert date-time variables into POSIXct/POSIXt format.}
\usage{
convert_datetime(
  dt_var,
  format = "ymd HM",
  check_ts_zero = FALSE,
  addtl_msg = NULL
)
}
\arguments{
\item{dt_var}{\cr
A vector containing the date-time values to be converted to the specified
format. The vector can be of any class (e.g., "character" or "POSIXct"). If
users have already pre-processed date-time variables into the expected
format (e.g., using \code{lubridate::ymd_hm()}) prior to running the function,
the function will simply return the original \code{dt_var}.}

\item{format}{(\code{character})\cr
Format to be used to parse \code{dt_var}, such as \code{"ymd HM"} (default),
\code{"ymd HMS"}, \code{"ymd"} etc. Will be passed to \code{lubridate::parse_date_time()}.

Multiple acceptable formats can be specified by providing a character vector,
e.g. \code{format = c("ymd HM", "ymd HMS", "ymd")}.

For more details, see \code{orders} input argument in
\link[lubridate]{parse_date_time}.}

\item{check_ts_zero}{(\code{logical})\cr
Whether to check for timestamps "00:00", which may represent missing times
for some variables. This is only relevant if \code{dtvar} is expected to have time
information, e.g., \code{format = c("ymd HM", "ymd HMS")}.}

\item{addtl_msg}{(\code{character})\cr
Additional warning message to be shown if any missing/invalid date-time
entries were found. If \code{addtl_msg = NULL}, the following message will be
shown by default:
"Please carefully consider how to deal with missing/invalid date-time entries
and perform any additional pre-processing prior to running the function
[\function_name] (e.g., impute missing dates/timestamps etc.)."

[\function_name]: R:\%5C\%5Cfunction_name\%5C}
}
\value{
(\code{POSIXct} | \code{POSIXt})\cr
Returns converted \code{dt_var} parsed according to specified date-time \code{format}.

Will also return the following warning messages (if applicable):
\enumerate{
\item Number (\%) of entries in \code{dt_var} that are missing (NA, "", " ")
\item Number (\%) of entries in \code{dt_var} that could not be parsed into specified
format
\item For formats containing timestamps: Number (\%) of entries in \code{dt_var} that
only contain date information
}

Any missing/invalid date-time entries will be returned as \code{NA} by this
function.
}
\description{
This function converts date-time variables into a user-specified format.
\code{convert_datetime} is called by other \code{Rgemini} functions to make sure that
any date-time inputs are in the expected format (typically \code{"ymd HM"}). It
can also be used independently of other \code{Rgemini} functions to clean up
date-times for analyses.
The function additionally returns warning messages about missing/invalid
date-time entries that can be useful for debugging.
}
\examples{
my_date_time <- c("2020-01-01 12:00", "2021-03-06 09:25")
convert_datetime(my_date_time, format = "ymd HM")
}
