% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_missing}
\alias{count_missing}
\title{Count missing}
\usage{
count_missing(x)
}
\arguments{
\item{x}{(\code{vector})}
}
\description{
This function checks the number of missingness in a vector, returns the results in the exact number (percentage)
}
