% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ntiles}
\alias{create_ntiles}
\title{Create N-tiles}
\usage{
create_ntiles(x, n)
}
\arguments{
\item{x}{(\code{vector})\cr
A vector of numeric values.}

\item{n}{(\code{integer})\cr
The number of bins to create. For example, for quartiles set \code{n=4}.}
}
\value{
(\code{factor})\cr
A factor with \code{n} levels representing the bins (ntiles) of the input values.
The factor returned will be the same length as \code{x}.
}
\description{
This function bins continuous variables into quantiles
(quartiles, deciles, etc.) of the user's choosing.
Note: The function will quit if any breakpoints are
duplicated (e.g. 1st quartile identical to 2nd quartile).
}
\examples{
set.seed(123)
values <- rnorm(100)
quartiles <- create_ntiles(values, 4)
deciles <- create_ntiles(values, 10)

}
