% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da21uid_statcan_v2021.R
\docType{data}
\name{da21uid_statcan_v2021}
\alias{da21uid_statcan_v2021}
\title{StatCan da21uid lookup}
\format{
A data table with columns:
\describe{
\item{da21uid}{Dissemination area IDs in Canada}
}
}
\usage{
da21uid_statcan_v2021
}
\description{
StatCan da21uid lookup
}
\keyword{datasets}
