% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_time_of_admission.R
\name{day_time_of_admission}
\alias{day_time_of_admission}
\title{Compute day & time of hospital admissions}
\usage{
day_time_of_admission(ipadmdad, dtvar1 = "admission_date_time")
}
\arguments{
\item{ipadmdad}{a data.table object equivalent of DRM table "ipadmdad". Table must contain two fields,
admission id and admission date and time}

\item{dtvar1}{a character string of label equivalent to "admission date time".
Its value must be in "yyyy-mm-dd hh:mm" format}
}
\value{
data.table with the same number of rows with input "ipadmdad" with addtional
derived character fields labelled as "day_of_admission_derived" and "time_of_admission_derived".
Possible values of these fields are "weekend" or "weekday" for the former, "daytime" or "nighttime" for the latter
}
\description{
\code{day_time_of_admission} returns whether the patient was admitted during day/night time and
weekday/weekend.
}
\details{
Day time is defined between 08:00 ~ 16:59, and 17:00 ~ 07:59 for Night time.
Weekday is Monday ~ Friday, weekend as Sunday and Saturday.

This classification might be useful as baseline admission characteristic in Epidemiological studies.
}
\section{Warning}{

NA values in returned data.table indicates either missing admission date and time or
its format is incorrect
}

\section{Note}{

Function takes required fields labels as input, this was intended not to hardcode variable names
within the function. When field labels change in database, there is no need to go through the lines to change
the hardcoded variable names
}

