% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_total_inpatient_cost.R
\name{derive_total_inpatient_cost}
\alias{derive_total_inpatient_cost}
\title{Derive total inpatient cost for each hospital stay.}
\usage{
derive_total_inpatient_cost(
  dbcon,
  cohort,
  reference_year = NA,
  hospital_level = FALSE
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{cohort}{(\code{data.frame} or \code{data.table})\cr
User specified cohort that's a restricted subset of all encounters in the DRM
table "ipadmdad" (see \href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}).
Must contain \code{genc_id} as the identifier, as well as admission_date_time and
discharge_date_time columns.}

\item{reference_year}{(\code{int})\cr
The year to which the derived costs should be adjusted to due to inflation.
As a default, reference year is decided to be the fiscal year of the latest
discharge in the provided cohort. For example, if the latest discharge is
on 2023-03-24, the reference year will be 2022 aligning with the fiscal year.}

\item{hospital_level}{(\code{logical})\cr
Logical indicating whether the user wants costs to be calculated using the
"Cost Per Standard Hospital Stay" at the hospital level rather than the
provincial level. Default is FALSE.}
}
\value{
This function returns a \code{data.table} containing:
\itemize{
\item \code{genc_id}: gemini encounter_id
\item \code{derived_total_inpatient_cost}: Derived inpatient cost for the current
encounter using the RIW method, without other adjustments.
\item \code{inflation_rate}: Inflation rate from discharge year to reference year.
\item 'derived_total_inpatient_cost_adjusted': cost adjusted to prices in the
provided reference year. If the reference year is not provided, then costs
are adjusted to the most recent fiscal year in the cohort.
}
}
\description{
\code{derive_total_inpatient_cost.R} derives the total cost of an encounter's
hospitalization based on Resource Intensity Weights (RIW), which represent
weighted costs relative to average inpatient cost. This function is currently
only able to derive inpatient costs for encounters with a methodology year
that's between 2017 and 2022.
}
\references{
include references to CIHI's CSHS table
}
