% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_index.R
\name{diagnoses_at_admission}
\alias{diagnoses_at_admission}
\title{Determine Diagnoses at Admission}
\usage{
diagnoses_at_admission(ipdiag, erdiag)
}
\arguments{
\item{ipdiag}{(\code{data.table} or \code{data.frame})\cr
\code{ipdiagnosis} table as defined in the
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.}

\item{erdiag}{(\code{data.table} or \code{data.frame})\cr
\code{erdiagnosis} table as defined in the
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
Typically, ER diagnoses should be included when deriving comorbidity in order
to increase sensitivity. However, in certain scenarios, users may choose to
only include IP diagnoses by specifying \code{erdiag = NULL}. This may be useful
when comparing cohorts with different rates of ER admissions.}
}
\value{
(\code{data.table})\cr
All encounters by \code{genc_id} with \code{diagnosis_type} and \code{diagnosis_code} for all diagnoses at admission.
}
\description{
See CIHI's \href{https://www.cihi.ca/sites/default/files/document/hospital-standardized-mortality-ratio-meth-notes-en.pdf}{HSMR Methodology}.

Specifically page 19 for inclusion/exclusion criteria.
}
