% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disability.R
\name{disability}
\alias{disability}
\title{Identify encounters with disability}
\usage{
disability(cohort, ipdiag, erdiag, component_wise = FALSE)
}
\arguments{
\item{cohort}{(\code{data.frame} or \code{data.table})
Cohort table with all relevant encounters of interest, where each row
corresponds to a single encounter. Must contain GEMINI Encounter ID
(\code{genc_id}).}

\item{ipdiag}{(\code{data.table})
\code{ipdiagnosis} table as defined in the \href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
This table must contain \code{genc_id} and \code{diagnosis_code} (as ICD-10-CA
alphanumeric code) in long format.}

\item{erdiag}{(\code{data.table})
\code{erdiagnosis} table as defined in the \href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
This table must contain \code{genc_id} and \code{er_diagnosis_code} (as ICD-10-CA
alphanumeric code) in long format.
Typically, ER diagnoses should be included when deriving disability in order
to increase sensitivity. However, in certain scenarios, users may choose to
only include IP diagnoses by specifying \code{erdiag = NULL}. This may be useful
when comparing cohorts with different rates of ER admissions.}

\item{component_wise}{(\code{logical})
If \code{component_wise == FALSE} (default), the function calculates a single
(global) disability flag indicating whether each \code{genc_id} was diagnosed with
\emph{any} disability.

If \code{component_wise == TRUE}, for each \code{genc_id} with a disability,
all identified disability diagnosis codes are returned, together with one of
the following 7 disability categories:
\itemize{
\item Physical disability - Congenital Anomalies
\item Physical disability - Musculoskeletal disorders
\item Physical disability - Neurological disorders
\item Physical disability - Permanent Injuries
\item Sensory disabilities - Hearing impairments
\item Sensory disabilities - Vision impairments
\item Developmental Disabilities
}}
}
\value{
\code{data.table}
If \code{component_wise == FALSE}, returns a table with all encounters identified
by the \code{cohort} table input and an additional derived field \code{disability}
(\code{logical}) indicating whether any diagnosis code for a disability was
identified. If a \code{genc_id} does not have any entry in the diagnosis table at
all, \code{disability = NA}. Note that this is very rare: If no additional
filtering was performed, >99.9\% of \code{genc_ids} should have an entry in the
\code{ipdiagnosis} (and >99.9\% of \code{genc_ids} that were admitted via ER should have
an entry in the \code{erdiagnosis} table).

If \code{component_wise == TRUE}, will only return \code{genc_ids} with a disability.
The output is returned in long format, where each row corresponds to a
disability diagnosis (\code{diagnosis_code}) and its corresponding
\code{disability_category} (\code{character}).
}
\description{
This function identifies whether or not an encounter had any ICD-10-CA
diagnosis code(s) indicating a physical, sensory, or intellectual/
developmental disability as defined by
\href{https://jamanetwork.com/journals/jamanetworkopen/fullarticle/2776018}{Brown et al., 2021}.

Broadly, this includes any diagnosis codes for conditions that are likely to
result in functional limitations and are considered to be chronic (see
Supplemental eTable 2 in \href{https://jamanetwork.com/journals/jamanetworkopen/fullarticle/2776018}{Brown et al., 2021}
for a full list of diagnosis codes). However, note that this function only
searches for relevant diagnosis codes at the encounter level and does not
check whether chronic conditions are coded consistently across different
encounters for a given patient.

By setting \code{component_wise} to \code{TRUE}, users can choose to return all
identified diagnosis codes for encounters with a disability, together with
the disability category each code was matched with (e.g., physical/sensory/
developmental disability - see below).
}
\section{Notes}{

This function does not differentiate between diagnosis types. That is, the
disability flags are derived based on all diagnosis codes that are provided
as input to this function. By default, users should include all diagnosis
types to identify disabilities. However, if users wish to include only
certain diagnosis types (e.g., type-M for most responsible discharge
diagnosis), the \code{ipdiag} and \code{erdiag} input tables should be filtered
accordingly based on \code{diagnosis_type} \emph{prior to running this function} (for
more details, see \href{https://www.cihi.ca/sites/default/files/document/diagnosis-type-definitions-en.pdf}{CIHI diagnosis type definitions}.
}

\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
  dbname = "db",
  host = "domain_name.ca",
  port = 1234,
  user = getPass("Enter user:"),
  password = getPass("password")
)

ipadmdad <- dbGetQuery(dbcon, "select * from admdad") \%>\% data.table()
ipdiagnosis <- dbGetQuery(dbcon, "select * from ipdiagnosis") \%>\% data.table()
erdiagnosis <- dbGetQuery(db, "select * from erdiagnosis") \%>\% data.table()

# including ER diagnosis codes
disability(cohort = ipadmdad, ipdiag = ipdiagnosis, erdiag = erdiagnosis)

# not including ER diagnosis codes
disability(cohort = ipadmdad, ipdiag = ipdiagnosis, erdiag = NULL)

# returning component-wise disability categories
disability(ipadmdad, ipdiagnosis, erdiagnosis, component_wise = TRUE)
}

}
\references{
Brown HK, et al. JAMA Netw Open, 2021. https://doi.org/10.1001/jamanetworkopen.2020.34993
}
