% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_admdad}
\alias{dummy_admdad}
\title{Generated simulated administrative data}
\usage{
dummy_admdad(id, admtime)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated in Rgemini version 3.0.0.
}
