% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_er}
\alias{dummy_er}
\title{Generate simulated ER data}
\usage{
dummy_er(
  nid = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  cohort = NULL,
  seed = NULL
)
}
\arguments{
\item{nid}{(\code{integer}) Number of unique encounter IDs to simulate. In this data table, each ID occurs once.}

\item{n_hospitals}{(\code{integer}) Number of hospitals in simulated dataset}

\item{time_period}{(\code{numeric}): Date range of data, by years or specific dates in either format:
("yyyy-mm-dd", "yyyy-mm-dd") or (yyyy, yyyy)}

\item{cohort}{(\code{data.frame}): Optional, data frame containing inpatient information
(encounter ID, hospital numbers, admission and discharge dates and times), resembling the GEMINI "admdad" table.}

\item{seed}{(\code{integer}) Optional, a number for setting the seed to get reproducible results.}
}
\value{
(\code{data.table}) A data.table object similar to the "er" table that contains the following fields:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{hospital_num} (\code{integer}): Hospital ID
\item \code{triage_date_time} (\code{character}): The date and time of triage
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "er" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.

This function will return one triage date time for each encounter ID.
}
\examples{
cohort <- dummy_ipadmdad()
dummy_er(cohort = cohort, seed = 1)
dummy_er(nid = 10, n_hospitals = 1, seed = 2)
}
