% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_ipadmdad}
\alias{dummy_ipadmdad}
\title{Simulate ipadmdad data}
\usage{
dummy_ipadmdad(n = 1000, n_hospitals = 10, time_period = c(2015, 2023))
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated in Rgemini version 3.0.0.
}
