% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_admdad_copy.R
\name{dummy_ipadmdad_copy}
\alias{dummy_ipadmdad_copy}
\title{Simulate ipadmdad data}
\usage{
dummy_ipadmdad_copy(
  n = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  seed = NULL
)
}
\arguments{
\item{n}{(\code{integer})\cr Total number of encounters (\code{genc_ids}) to be
simulated.}

\item{n_hospitals}{(\code{integer})\cr
Number of hospitals to be simulated. Total number of \code{genc_ids} will be split
up pseudo-randomly between hospitals to ensure roughly equal sample size at
each hospital.}

\item{time_period}{(\code{numeric})\cr
A numeric vector containing the time period, specified as fiscal years
(starting in April each year). For example, \code{c(2015, 2019)} generates data
from 2015-04-01 to 2020-03-31.}
}
\value{
(\code{data.frame})\cr A data.frame object similar to the "ipadmdad" table
containing the following fields:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{hospital_num} (\code{integer}): Hospital ID
\item \code{admission_date_time} (\code{character}): Date-time of admission in YYYY-MM-DD HH:MM format
\item \code{discharge_date_time} (\code{character}): Date-time of discharge in YYYY-MM-DD HH:MM format
\item \code{age} (\code{integer}): Patient age
\item \code{gender} (\code{character}): Patient gender (F/M/O for Female/Male/Other)
\item \code{discharge_disposition} (\code{integer}): All valid categories according to DAD
abstracting manual 2022-2023
\itemize{
\item 4: Home with Support/Referral
\item 5: Private Home
\item 8: Cadaveric Donor (does not exist in GEMINI data)
\item 9: Stillbirth (does not exist in GEMINI data)
\item 10: Transfer to Inpatient Care
\item 20: Transfer to ED and Ambulatory Care
\item 30: Transfer to Residential Care
\item 40: Transfer to Group/Supportive Living
\item 90: Transfer to Correctional Facility
\item 61: Absent Without Leave (AWOL)
\item 62: Left Against Medical Advice (LAMA)
\item 65: Did not Return from Pass/Leave
\item 66: Died While on Pass/Leave
\item 67: Suicide out of Facility (does not exist in GEMINI data)
\item 72: Died in Facility
\item 73: Medical Assistance in Dying (MAID)
\item 74: Suicide in Facility
}
\item \code{alc_service_transfer_flag} (\code{character}): Variable indicating whether
patient was transferred to an alternate level of care (ALC) during their
hospital stay. Coding is messy and varies across sites. Possible values are:
\itemize{
\item Missing: \code{NA}, \code{""}
\item True: \code{"TRUE"/"true"/"T"}, \code{"y"/"Y"}, \code{"1"/"99"}, \code{"ALC"}
\item False: \code{"FALSE"/"false"}, \code{"N"}, \code{"0"}, \code{"non-ALC"}
Some entries with missing \code{alc_service_transfer_flag} can be inferred based
on value of \code{number_of_alc_days} (see below)
}
\item \code{number_of_alc_days} (\code{integer}): Number of days spent in ALC (rounded to
nearest integer). If \code{number_of_alc_days = 0}, no ALC occurred;
if \code{number_of_alc_days > 0}, ALC occurred.
Note that days spent in ALC should usually be < length of
stay. However, due to the fact that ALC days are rounded up, it's possible
for \code{number_of_alc_days} to be larger than \code{los_days_derived}.
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "ipadmdad" table (see details in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}). It is a copy of the Rgemini function \code{dummy_ipadmdad}.

The simulated encounter-level variables that are returned by this function
are currently: Admission date-time, discharge date-time, age, gender,
discharge disposition, transfer to an alternate level of care (ALC), and ALC
days. The distribution of these simulated variables roughly mimics the real
distribution of each variable observed in the GIM cohort from 2015-2022.
Admission date-time is simulated in conjunction with discharge date-time to
mimic realistic length of stay. All other variables are simulated
independently of each other, i.e., there is no correlation between age,
gender, discharge disposition etc. that may exist in the real data. One
exception to this is \code{number_of_alc_days}, which is only > 0 for entries
where \code{alc_service_transfer_flag == TRUE} and the length of ALC is capped at
the total length of stay.

The function simulates patient populations that differ across hospitals. That
is, patient characteristics are simulated separately for each hospital, with
a different, randomly drawn distribution mean (i.e., random intercepts).
However, the degree of hospital-level variation simulated by this function
is arbitrary and does not reflect true differences between hospitals in the
real GEMINI dataset.
}
\examples{
# Simulate 10,000 encounters from 10 hospitals for fiscal years 2018-2020.
ipadmdad <- dummy_ipadmdad(n = 10000, n_hospitals = 10, time_period = c(2018, 2020))

}
