% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/dummy_data.R
\docType{data}
\name{dummy_ipdiagnosis}
\alias{dummy_ipdiagnosis}
\title{Dummy ipdiagnosis data}
\description{
This mimics the GEMINI ipdiagnosis table. And is generated with the \code{dummy_diag} function.
See the \href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}
for details.
}
\keyword{internal}
