% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_ipscu}
\alias{dummy_ipscu}
\title{Generate simulated ipscu data}
\usage{
dummy_ipscu(
  nid = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  cohort = NULL,
  seed = NULL
)
}
\arguments{
\item{nid}{(\code{integer})\cr Number of unique encounter IDs to simulate.
Encounter IDs may repeat, resulting in a data table with more rows than \code{nid}.}

\item{n_hospitals}{(\code{integer})\cr Number of hospitals in simulated dataset}

\item{time_period}{(\code{numeric})\cr Date range of data, by years or specific dates in either format:
("yyyy-mm-dd", "yyyy-mm-dd") or (yyyy, yyyy)}

\item{cohort}{(\code{data.frame})\cr Optional, data table containing inpatient information
(encounter ID, hospital numbers, admission and discharge dates and times), resembling the GEMINI "admdad" table.}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results}
}
\value{
(\code{data.table})\cr A data.table object similar to the "ipscu" table that contains the following fields:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{hospital_num} (\code{integer}): Hospital ID
\item \code{scu_admit_date_time} (\code{character}): Date and time of SCU admission in YYYY-MM-DD HH:MM format
\item \code{scu_discharge_date_time} (\code{character}): Date and time of SCU admission in YYYY-MM-DD HH:MM format
\item \code{icu_flag} (\code{logical}): Flag specifying whether the encounter was admitted to the ICU or not.
This refers to to SCU unit numbers excluding the step down units of 90, 93, and 95.
\item \code{scu_unit_number}(\code{integer}): Code identifying the type of special care unit where the patient receives
critical care, according to DAD abstracting manual 2025-2026
\itemize{
\item 10: Medical Intensive Care Nursing Unit
\item 20: Surgical Intensive Care Nursing Unit
\item 25: Trauma Intensive Care Nursing Unit
\item 30: Combined Medical/Surgical Intensive Care Nursing Unit
\item 35: Burn Intensive Care Nursing Unit
\item 40: Cardiac Intensive Care Nursing Unit Surgery (CCU)
\item 45: Coronary Intensive Care Nursing Unit Medical (CCU)
\item 50: Neonatal Intensive Care Nursing Unit Undifferentiated/General
\item 60: Neurosurgery Intensive Care Nursing Unit
\item 80: Respirology Intensive Care Nursing Unit
\item 90: Step-Down Medical Unit
\item 93: Combined Medical/Surgical Step-Down Unit
\item 95: Step-Down Surgical Unit
}
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "ipscu" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
}
\examples{
dummy_ipscu(nid = 100, n_hospitals = 10, time_period = c(2015, 2023))
dummy_ipscu(nid = 11, n_hospitals = 1, time_period = c("2020-01-01", "2021-01-01"))

}
