% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_lab}
\alias{dummy_lab}
\title{Generated simulated lab data}
\usage{
dummy_lab(id, omop, value, unit, mintime)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated in Rgemini version 3.0.0.
}
