% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_locality}
\alias{dummy_locality}
\title{Generate simulated locality variables data}
\usage{
dummy_locality(
  dbcon = NULL,
  nid = 1000,
  n_hospitals = 10,
  cohort = NULL,
  da21uid = NULL,
  seed = NULL
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A connection to the GEMINI database, used to access the 2021 Canadian census dissemination codes.
It is required when \code{da21uid} is missing.}

\item{nid}{(\code{integer})\cr Number of unique encounter IDs to simulate. In this data table, each ID occurs once.
It is optional when \code{cohort} is provided.}

\item{n_hospitals}{(\code{integer})\cr Number of hospitals in simulated dataset.
IT is optional when \code{cohort} is provided.}

\item{cohort}{(\code{data.frame | data.table}) Optional, an existing data table similar to \code{admdad} in GEMINI
with at least the following columns:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{hospital_num} (\code{integer}): Hospital ID
If \code{cohort} is provided, \code{nid} and \code{n_hospital} inputs are not used.
}}

\item{da21uid}{(\code{integer} | \code{vector})\cr Allows the user to customize which location ID to include in the output.
It is required when \code{dbcon} is missing. It can be an integer or an integer vector.}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results.}
}
\value{
(\code{data.table})\cr
A data.table object similar to the "locality_variables" table that contains the following fields
(if \code{cohort} is included, these are in addition to any columns in \code{cohort}):
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{da21uid} (\code{integer}): Dissemination area ID based on 2021 Canadian census data using PCCF Version 8A
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "locality-variables" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.

This data table will simulate dissemination area IDs based on Canadian census data.
}
