% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_physicians}
\alias{dummy_physicians}
\title{Generate simulated physicians data}
\usage{
dummy_physicians(nid = 1000, n_hospitals = 10, cohort = NULL, seed = NULL)
}
\arguments{
\item{nid}{(\code{integer})\cr Number of unique encounter IDs to simulate.
Optional if \code{cohort} is provided.}

\item{n_hospitals}{(\code{integer})\cr Number of hospitals in simulated dataset.
Optional if \code{cohort} is provided.}

\item{cohort}{(\verb{data.frame or data.table}) Optional, an existing data table or data frame
similar to \code{admdad} in GEMINI with at least the following columns:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID; integers starting from 1
\item \code{hospital_num} (\code{integer}): Mock hospital ID; integers starting from 1
If \code{cohort} is provided, \code{nid} and \code{n_hospitals} inputs are not used.
}}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results.}
}
\value{
(\code{data.table})\cr A data.table object similar to the "physicians" table that contains the
following fields:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID; integers starting from 1
\item \code{hospital_num} (\code{integer}): Mock hospital ID number; integers starting from 1
\item \code{admitting_physician_gim} (\code{logical}): Whether the admitting physician attends a general medicine ward
\item \code{discharging_physician_gim} (\code{logical}): Whether the discharging physician attends a general medicine ward
\item \code{adm_phy_cpso_mapped} (\code{integer}): Unique hash of admitting physician CPSO Number
\item \code{mrp_cpso_mapped} (\code{integer}): Unique hash of most responsible physician (MRP) CPSO Number
\item \code{dis_phy_cpso_mapped} (\code{integer}): Unique hash of discharging physician CPSO Number
}
}
\description{
This function creates a synthetic dataset with a subset of variables that
are contained in the GEMINI "physicians" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
}
\examples{
dummy_physicians(nid = 1000, n_hospitals = 10, seed = 1)
dummy_physicians(cohort = dummy_ipadmdad(), seed = 2)

}
