% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_radiology}
\alias{dummy_radiology}
\title{Generate simulated radiology data}
\usage{
dummy_radiology(
  nid = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  cohort = NULL,
  seed = NULL
)
}
\arguments{
\item{nid}{(\code{integer})\cr Number of unique encounter IDs to simulate.
Encounter IDs may repeat to simulate multiple radiology tests,
resulting in a data table with more rows than \code{nid}.}

\item{time_period}{(\code{vector})\cr A numeric or character vector containing the data range of the data
by years or specific dates in either format: ("yyyy-mm-dd", "yyyy-mm-dd") or (yyyy, yyyy)
The start date and end date will be (yyyy-01-01 and yyyy-12-31) if (yyyy, yyyy)
is the date range format provided. Optional when \code{cohort} is provided.}

\item{cohort}{(\code{data.frame | data.table})\cr Optional, data frame with the following columns:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{hospital_num} (\code{integer}): Hospital ID
\item \code{admission_date_time} (\code{character}): Date and time of IP admission in YYYY-MM-DD HH:MM format
\item \code{discharge_date_time} (\code{character}): Date and time of IP discharge in YYYY-MM-DD HH:MM format.
When \code{cohort} is not NULL, \code{nid}, \code{n_hospitals}, and \code{time_period} are ignored.
}}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results.}

\item{n_hospitals(`integer`)\cr}{The number of hospitals to simulate.
It is optional if \code{cohort} is provided.}
}
\value{
(\code{data.table})\cr A \code{data.table} object similar that contains the following fields:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{modality_mapped} (\code{character}): Imaging modality: either MRI, CT, or Ultrasound.
\item \code{ordered_date_time} (\code{character}): The date and time the radiology test was ordered
\item \code{performed_date_time} (\code{character}): The date and time the radiology test was performed
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "radiology" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.

This function only simulates modalities used in Our Practice Report (OPR) - CT, MRI,
Ultrasound. It does not cover all modalities seen in the actual "radiology" data table.
}
\examples{
cohort <- dummy_ipadmdad()
dummy_radiology(cohort = cohort)
dummy_radiology(n = 1000, n_hospitals = 10, time_period = c(2020, 2023))

}
