% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_transfusion}
\alias{dummy_transfusion}
\title{Generate simulated transfusion data}
\usage{
dummy_transfusion(
  nid = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  cohort = NULL,
  blood_product_list = NULL,
  seed = NULL
)
}
\arguments{
\item{nid}{(\code{integer})\cr The number of unique mock encounter IDs to simulate.
Encounter IDs may repeat, resulting in a data table with more rows than \code{nid}. Optional if \code{cohort} is provided.}

\item{n_hospitals}{(\code{integer})\cr The number of hospitals to simulate, optional if \code{cohort} is provided.}

\item{time_period}{(\code{vector})\cr A numeric or character vector containing the data range of the data
by years or specific dates in either format: ("yyyy-mm-dd", "yyyy-mm-dd") or (yyyy, yyyy).
The start date and end date will be (yyyy-01-01 and yyyy-12-31) if (yyyy, yyyy)
is the date range format provided. Not used when \code{cohort} is provided.}

\item{cohort}{(\verb{data.frame or data.table})\cr Optional, a data frame or data table with columns:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID number
\item \code{hospital_num} (\code{integer}): Mock hospital ID number
\item \code{admission_date_time} (\code{character}): Date and time of IP admission in YYYY-MM-DD HH:MM format.
\item \code{discharge_date_time} (\code{character}): Date and time of IP discharge in YYYY-MM-DD HH:MM format.
When \code{cohort} is not NULL, \code{nid}, \code{n_hospitals}, and \code{time_period} are ignored.
}}

\item{blood_product_list}{(\code{character})\cr Either a string or a character vector
to sample for the variable \code{blood_product_mapped_omop}.
Items must be real blood product OMOP codes or it will not be used and a warning will be raised.}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results.}
}
\value{
(\code{data.table})\cr A data.table object similar to the "transfusion" table with the following fields:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID number; integers starting from 1 or from \code{cohort}
\item \code{hospital_num} (\code{integer}): Mock hospital ID number; integers starting from 1 or from \code{cohort}
\item \code{issue_date_time} (\code{character}): The date and time the transfusion was issued, in the format ("yy-mm-dd hh:mm")
\item \code{blood_product_mapped_omop} (\code{character}): Blood product name mapped by GEMINI following international standard.
\item \code{blood_product_raw} (\code{character}): Type of blood product or component transfused as reported by hospital.
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "transfusion" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
}
\examples{
dummy_transfusion(nid = 1000, n_hospitals = 30, seed = 1)
dummy_transfusion(cohort = dummy_ipadmdad())
dummy_transfusion(nid = 100, n_hospitals = 1, blood_product_list = c("0", "35605159", "35615187"))

}
