% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_index.R
\name{elixhauser_comorbidity_index}
\alias{elixhauser_comorbidity_index}
\title{Compute Elixhauser comorbidity score}
\usage{
elixhauser_comorbidity_index(
  ipdiag,
  erdiag,
  at_admission = TRUE,
  raw_comorbidities = FALSE
)
}
\arguments{
\item{ipdiag}{(\code{data.table} or \code{data.frame})\cr
\code{ipdiagnosis} table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.}

\item{erdiag}{(\code{data.table} or \code{data.frame})\cr
\code{erdiagnosis} table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.
Typically, ER diagnoses should be included when deriving comorbidity in order
to increase sensitivity. However, in certain scenarios, users may choose to
only include IP diagnoses by specifying \code{erdiag = NULL}. This may be useful
when comparing cohorts with different rates of ER admissions.}

\item{at_admission}{(\code{logical})\cr
Whether to calculate the score for all comorbidities or for only pre-admit comorbidities.}

\item{raw_comorbidities}{(\code{logical})\cr
Whether to output a \code{data.table} of raw comorbidities as opposed to pre-computed scores.}
}
\value{
(\code{data.table})\cr
By default, for each encounter, outputs the comorbidity score. If \code{raw_comorbidities} is \code{TRUE},
outputs a wide \code{data.table} with a column for each comorbidity for each encounter.
}
\description{
Based on the methodology from \href{https://pubmed.ncbi.nlm.nih.gov/19433995/}{Elixhauser A et al. 1998}
using ICD-10 codes as the mapping algorithm,
implemented in the R \href{https://ellessenne.github.io/comorbidity/index.html}{comorbidity} package
by \href{https://doi.org/10.21105/joss.00648}{Gasparini, 2018}.

Can compute either Elixhauser score at admission based on
\href{https://www.cihi.ca/sites/default/files/document/hospital-standardized-mortality-ratio-meth-notes-en.pdf}{HSMR Methodology}
or for the entire patient encounter. The default is to compute comorbidity at admission.
}
\details{
A hierarchy of comorbidities is used when calculating the cormorbidity score, but not when outputting \code{raw_comorbidites}.
This affects comorbidities present in a patient with different degrees of severity. See documentation for
\code{\link[comorbidity:comorbidity]{comorbidity::comorbidity()}} for details.
}
\references{
\itemize{
\item{Elixhauser A, et al. Med Care, 1998. https://doi.org/10.1097/MLR.0b013e31819432e5}
\item{Gasparini A. JOSS, 2018. https://doi.org/10.21105/joss.00648}
}
}
