% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_of_care.R
\name{episodes_of_care}
\alias{episodes_of_care}
\title{Identify episodes of care}
\usage{
episodes_of_care(dbcon, restricted_cohort = NULL)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{restricted_cohort}{(\code{data.table} | \code{data.frame})\cr
User specified cohort that is a restricted subset of all encounters in DRM table "ipadmdad" (see
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}).
Must contain \code{genc_id} as the identifier. Default is \code{Null}, which loads the entire "ipadmdad"
table in the user-provided database (recommended approach).}
}
\value{
(\code{data.frame})\cr
This function returns a \code{data.table} excluding records with invalid \code{patient_id_hashed}, \code{age < 18}, or
\code{admit_category} of cadaveric donors in line with CIHI guidelines.
Nine columns are returned: \code{genc_id}, \code{patient_id_hashed}, \code{time_to_next_admission} (\code{numeric}),
\code{time_since_last_admission} (\code{numeric}), \code{AT_in_coded} (\code{TRUE}/\code{FALSE}), \code{AT_out_coded} (\code{TRUE}/\code{FALSE}),
\code{AT_in_occurred} (\code{TRUE}/\code{FALSE}/\code{NA}), \code{AT_out_occurred} (\code{TRUE}/\code{FALSE}/\code{NA}), \code{epicare} (\code{numeric})
}
\description{
The \code{episodes_of_care()} function groups encounters that are linked via transfers into a single episode of
care. The derived episodes of care are used in the \code{readmission()} function to avoid counting transfers as
readmissions. Therefore, the criteria for transfers used in this function are based on
\href{https://www.cihi.ca/en/indicators/all-patients-readmitted-to-hospital}{CIHI guidelines for readmission calculations}.
However, the \code{episodes_of_care} function can also be used independently to compute episodes
of care outside of readmission.

Transfer information is obtained from the GEMINI database table \code{lookup_transfer} (see
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI database schema}).
By default, the function queries all encounters in the \code{admdad} table to ensure that any encounters that may be
linked via transfers are accurately grouped into the same episode of care.

Note: Users do not typically need to calculate episodes of care themselves, but rather should use the derived
\code{epicares} variable in the GEMINI database, if available. This is due to the fact that the derived variable
is calculated based on all available GEMINI encounters, and therefore, provides the most accurate identification
of episodes of care. By contrast, datacuts or cohorts that have been pre-filtered may miss certain encounters,
resulting in a loss of information. However, in certain situations, users may still find this function useful 1)
to understand how the derived variable is calculated and 2) in case users need to re-calculate episodes of care
or readmission rates for a restricted cohort (see below).
}
\details{
An episode of care refers to all contiguous inpatient hospitalizations admitted to any medical or intensive care
service within GEMINI. Episodes involving inter-facility transfers are linked regardless of diagnosis. An acute
care transfer is assumed to have occurred if either of the following criteria are met (see
\href{https://www.cihi.ca/en/indicators/all-patients-readmitted-to-hospital}{CIHI guidelines}:
\enumerate{
\item An admission occurs within 7 hours after discharge, regardless of whether the transfer is coded by hospitals.
OR 2) An admission occurs within 7-12 hours after discharge, and at least one hospital has coded the transfer.
}

Acute transfers that are coded by hospitals (\code{AT_in_coded} and \code{AT_out_coded}) are defined by the DAD fields
\code{institution_from} and \code{institution_to}:
\code{AT_in_coded}/\code{AT_out_coded} is \code{TRUE}, when '\code{institution_from == "AT"} or '\code{institution_to == "AT"}. All
remaining entries are set to \code{FALSE}.

Acute transfers that actually occurred (\code{AT_in_occurred} and \code{AT_out_occurred}) are defined as follows:
\code{AT_in_occurred}/\code{AT_out_occurred} is \code{TRUE}, when admission is within 7 hrs of discharge regardless of transfer
coding, or, admission is within 7-12hrs of discharge and >=1 hospital coded the transfer.
\code{AT_in_occurred}/\code{AT_out_occurred} is \code{NA}, when admission time since previous discharge is unknown and=1 hospital coded the transfer. This is because it cannot be determined if the transfer actually took place ornot.AT_in_occurred/AT_out_occurred is FALSE, for all remaining entries.

Each episode of care (\code{epicare}) is defined by linked transfers identified based on \code{AT_in_occurred}/
\code{AT_out_occurred}. A unique numeric ID is assigned to each unique episode of care. That is, contiguous encounters
linked by transfers will have the same \code{epicare} number.
}
\note{
We recommend that users run this function on the entire GEMINI data set (the default setting) to ensure no
transfers to/from are omitted.
Please review the function documentation & vignette carefully. A comprehensive understanding of the following
concepts is needed for proper use of the function: transfer coded vs. transfer occurred, episode of care definition.
}
\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
                        dbname = "db",
                        host = "domain_name.ca",
                        port = 1234,
                        user = getPass("Enter user:"),
                        password = getPass("password"))

epicare_table <- episodes_of_care(dbcon)
}

}
\references{
\href{https://www.cihi.ca/en/indicators/all-patients-readmitted-to-hospital}{CIHI readmission guidelines}
}
\seealso{
\code{vignette("epicare_and_readmission", package = "Rgemini")}
}
