% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_var_str}
\alias{fix_var_str}
\title{Fix variable strings}
\usage{
fix_var_str(str)
}
\arguments{
\item{str}{(\code{character})
Character string to be cleaned up}
}
\description{
Removes any "_" from variable names and applies title case capitalization.
}
