% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{gemini_cache_context}
\alias{gemini_cache_context}
\title{Cache a Gemini context}
\usage{
gemini_cache_context(
  context,
  verbose = FALSE,
  api_key = getOption("gemini_api_key")
)
}
\arguments{
\item{context}{A `gemini_context` object.}

\item{verbose}{Logical; if `TRUE`, prints additional logging information.}

\item{api_key}{The API key for authentication (default is taken from `getOption("gemini_api_key")`).}
}
\value{
The updated `gemini_context` object with caching details.
}
\description{
Caches a Gemini context to the API.
}
