% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gemini_colors}
\alias{gemini_colors}
\title{GEMINI colors}
\usage{
gemini_colors(palette = "")
}
\arguments{
\item{palette}{\subsection{(\code{character})\cr
Name of color palette. Has to be one of the following:}{
\itemize{
\item 
}
}}
}
\description{
Defines various color palettes for plotting purposes.
Color palettes can be explored using \code{scales::show_col(gemini_colors())}
}
