% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgemini.R
\name{gemini_content}
\alias{gemini_content}
\title{Extract content from a more detailed run_gemini response}
\usage{
gemini_content(result)
}
\description{
Extract content from a more detailed run_gemini response
}
