% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{gemini_context}
\alias{gemini_context}
\title{Create a Gemini context}
\usage{
gemini_context(
  prompt = NULL,
  model = "gemini-1.5-flash-001",
  media = NULL,
  ttl_sec = 60 * 60,
  role = "user",
  do_cache = TRUE,
  api_key = getOption("gemini_api_key")
)
}
\arguments{
\item{prompt}{The input prompt for the model.}

\item{model}{The model to use (default: `"gemini-1.5-flash-001"`).}

\item{media}{Optional media input.}

\item{ttl_sec}{Time-to-live for cache in seconds (default: `60*60`).}

\item{role}{The role of the request (default: `"user"`).}

\item{do_cache}{Logical; if `TRUE`, attempts to cache the context.}

\item{api_key}{The API key for authentication (default is taken from `getOption("gemini_api_key")`).}
}
\value{
A `gemini_context` object.
}
\description{
Initializes a Gemini context object with caching options.
}
