% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{gemini_delete_all_context_caches}
\alias{gemini_delete_all_context_caches}
\title{Delete all cached contexts}
\usage{
gemini_delete_all_context_caches(api_key = getOption("gemini_api_key"))
}
\arguments{
\item{api_key}{The API key for authentication (default is taken from `getOption("gemini_api_key")`).}
}
\value{
The number of caches deleted.
}
\description{
Deletes all cached contexts from the Gemini API.
}
