% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{gemini_delete_context_cache}
\alias{gemini_delete_context_cache}
\title{Delete a cached context}
\usage{
gemini_delete_context_cache(
  context = NULL,
  cache_name = context$cache_name,
  api_key = getOption("gemini_api_key"),
  verbose = FALSE
)
}
\arguments{
\item{context}{An optional context object containing the cache name.}

\item{cache_name}{The name of the cache to delete (default is extracted from `context`).}

\item{api_key}{The API key for authentication (default is taken from `getOption("gemini_api_key")`).}

\item{verbose}{Logical; if TRUE, prints messages about the deletion process.}
}
\value{
Returns `TRUE` if deletion is successful, `FALSE` otherwise. If `context` is provided, returns the updated context.
}
\description{
Deletes a cached context from the Gemini API.
}
