% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{gemini_list_context_caches}
\alias{gemini_list_context_caches}
\title{List all cached contexts}
\usage{
gemini_list_context_caches(api_key = getOption("gemini_api_key"))
}
\arguments{
\item{api_key}{The API key for authentication (default is taken from `getOption("gemini_api_key")`).}
}
\value{
A data frame with cached context details, or `NULL` if no caches are found.
}
\description{
Retrieves a list of all cached contexts from the Gemini API.
}
