% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_upload.R
\name{gemini_media_upload}
\alias{gemini_media_upload}
\title{Upload a document or image to Gemini API.}
\usage{
gemini_media_upload(
  file_path,
  mime_type = NULL,
  display_name = NULL,
  api_key = getOption("gemini_api_key"),
  verbose = FALSE
)
}
\description{
Returned object or a list of such objects can then be passed to gemini_run.
 Media should be available for around an hour.
}
