% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_theme.R
\name{gemini_theme}
\alias{gemini_theme}
\title{Gemini base theme}
\usage{
gemini_theme(
  base.size = 14,
  base.family = "sans",
  base.colour = "steelblue",
  bg.colour = "lightgray"
)
}
\arguments{
\item{base.size}{}

\item{base.family}{}

\item{base.color}{}

\item{bg.color}{}
}
\description{
Gemini base theme
}
