% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{gemini_update_context_cache}
\alias{gemini_update_context_cache}
\title{Update cached context expiration}
\usage{
gemini_update_context_cache(
  context,
  ttl_sec = context$ttl_sec,
  only_if_worked = TRUE,
  only_if_expired = TRUE,
  expired_sec_margin = 10,
  api_key = getOption("gemini_api_key")
)
}
\arguments{
\item{context}{A `gemini_context` object.}

\item{ttl_sec}{New TTL value (default: existing TTL).}

\item{only_if_worked}{Logical; if `TRUE`, only updates if caching previously worked.}

\item{only_if_expired}{Logical; if `TRUE`, only updates if cache is expired.}

\item{expired_sec_margin}{Margin in seconds before cache is considered expired (default: `10`).}

\item{api_key}{The API key for authentication (default is taken from `getOption("gemini_api_key")`).}
}
\value{
The updated `gemini_context` object.
}
\description{
Updates the TTL of a cached Gemini context.
}
