% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not In}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{(\code{vector} or \code{NULL})
The values to be matched. Long vectors are supported.}

\item{y}{(\code{vector} or \code{NULL})
The values to be matched against. Long vectors are not supported.}
}
\value{
(\code{logical})
A vector of logical values the same length of \code{x}, for which each
value represents whether that particular element in \code{x} exists in \code{y}.
}
\description{
Infix function for the opposite of \code{\%in\%}.
}
\examples{
c("x", "y") \%ni\% "x"

}
