% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalizations_last_n_days.R
\name{hospitalizations_last_n_days}
\alias{hospitalizations_last_n_days}
\title{Compute number of previous hospitalizations for an encounter in a
given window.}
\usage{
hospitalizations_last_n_days(
  cohort,
  n_days = 30,
  admit_dt = "admission_date_time",
  discharge_dt = "discharge_date_time"
)
}
\arguments{
\item{cohort}{(\code{data.table}, \code{data.frame})\cr Cohort table with all relevant
encounters of interest, where each row corresponds to a single encounter.
Must contain the following columns:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{patient_id_hashed} (\code{Varchar}): Unique hash of patient's health card number
\item \code{admission_date_time} (\code{character}): The date and time that the patient was officially registered as an inpatient for the current hospital admission. Must be in YYYY-MM-DD HH:MM.
\item \code{discharge_date_time} (\code{character}): The date and time when the patient was formally discharged from the current hospital admission. Must be in YYYY-MM-DD HH:MM.
}}

\item{n_days}{(\code{integer})\cr Window for calculating the number of
hospitalizations. n_days is set to 30 days by default.}

\item{admit_dt}{(\code{character})\cr Admission date-time Variable representing
admission date-time. Set to admission_date_time by default.}

\item{discharge_dt}{Discharge date-time Variable representing discharge
date-time. Set to discharge_date_time by default.}
}
\value{
(\code{data.table})\cr data.table containing the same number of rows as
input \code{cohort} table. This data.table contains the genc_id, along with
\code{n_hospitalizations}, which represents the number of previous
hospitalizations in the given window. The values in \code{n_hospitalizations}
will be 0, 1, or 2+.
}
\description{
Computes the number of hospitalizations an encounter has had
over a specified time period, using admission & discharge date-times from
the CIHI Discharge Abstract Database (DAD).
}
\details{
This function uses DAD fields Admission date-time (Group 04 Fields
01/02), Discharge date-time (Group 05 Fields 01/02) along with
patient_id_hashed to calculate the number of previous hospitalizations an
encounter has had in a specified window.
}
\examples{
# Default n_days of 30:
\dontrun{
hospitalizations_last_n_days(cohort)
}

# User-input window of 182 days (6 months)
\dontrun{
hospitalizations_last_n_days(cohort, n_days = 182)
}


}
