% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_hospital_mortality.R
\name{in_hospital_mortality}
\alias{in_hospital_mortality}
\title{Compute in-hospital mortality using CIHI DAD}
\usage{
in_hospital_mortality(ipadmdad, suicide = FALSE)
}
\arguments{
\item{ipadmdad}{a data.table object equivalent of DRM table "ipadmdad".
Table must contain two fields:
admission id and discharge disposition}

\item{suicide}{a TRUE/FALSE argument indicating whether suicides in facility
should be factored in the calculation.}

\item{id}{a character string of label for admission id
in "ipadmdad" table}

\item{dischvar}{a character string of label equivalent to
CIHI-defined "discharge disposition".}
}
\value{
data.table with the same number of rows with input "ipadmdad" with additional
derived boolean field labelled as
"in_hospital_mortality_derived". Possible values are TRUE, FALSE or NA
}
\description{
\code{in_hospital_mortality} returns whether patient has deceased
in hospital based on the CIHI DAD field
}
\details{
This function takes the CIHI DAD "discharge disposition"
field (Group 05, Field 05 As of 19/20 CIHI DAD) and generate
boolean field indicating whether the patient has deceased during hospital stay.

The current version of the function captures the four possible
values indicating patient has deceased in hospital:

\itemize{
\item{7 :  }{Died (Deprecated post 2017)}
\item{72 : }{Died in Facility}
\item{73 : }{Medical Assistance in Dying}
\item{74 : }{Suicide in Facility (Excluded in calculation by default; can be included in argument)}
}
}
\section{Warning}{

NA values in returned data.table indicates missing
discharge disposition field values.
}

