% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlaps.R
\name{laps_assign_test}
\alias{laps_assign_test}
\title{Assign score to LAPS component}
\usage{
laps_assign_test(x, breaks, points)
}
\arguments{
\item{x}{(\code{character})\cr
A vector of result values for a particular lab test.}

\item{breaks}{(\code{numeric})\cr
A vector of intervals for a particular test which correspond to a different number of points.}

\item{points}{(\code{numeric})\cr
A vector of scores for each interval in \code{breaks}.}
}
\value{
(\code{numeric})\cr
A vector of scores for each lab result value provided in \code{x}.
}
\description{
Assign score to LAPS component
}
\examples{
x <- c("0.3", "0.7")
breaks <- c(0, 0.2, 0.4, 0.8, Inf)
points <- c(1, 2, 3, 4)

laps_assign_test(x, breaks, points)

}
