% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{length_of_stay}
\alias{length_of_stay}
\title{Compute hospital Length of Stay  (LoS)}
\usage{
length_of_stay(
  ipadmdad,
  adm_dtvar = "admission_date_time",
  dis_dtvar = "discharge_date_time"
)
}
\arguments{
\item{ipadmdad}{(\code{data.frame} or \code{data.table})\cr
Table with all relevant encounters of interest from DRM table "ipadmdad" (see
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}).
Must contain three fields: encounter ID (\code{genc_id}), and two date-time
variables corresponding to admission and discharge (typically
\code{admission_date_time} and \code{discharge_date_time}). Date-time variables must be
in "yyyy-mm-dd hh:mm" format.}

\item{adm_dtvar}{(\code{character})\cr
Character string defining the column name for admission date-time (typically
"admission_date_time").}

\item{dis_dtvar}{(\code{character})\cr
Character string defining the column name for discharge date-time (typically
"discharge_date_time").}
}
\value{
data.table with the same number of rows as input "ipadmdad", with additional
derived numeric fields labelled as "los_hrs_derived" and "los_days_derived".
}
\description{
\code{length_of_stay} returns Length of Stay for hospital admissions in unit of
hours and days.
}
\details{
Length of Stay is defined as the duration of hospital in-patient stay,
calculated as \code{discharge_date_time - admission_date_time}.
It is a clinical outcome variable in various research studies
and one of several indicators of quality of care in GeMQIN reports.
}
\section{Warning}{

NA values in returned data.table indicates either missing or invalid
admission/discharge date-time.
}

\section{Note}{

The function takes two optional input arguments defining the admission and
discharge date-time variables of interest (by default \code{admission_date_time}
and \code{discharge_date_time}).
}

