% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lunique}
\alias{lunique}
\title{Returns number of unique values}
\usage{
lunique(x)
}
\arguments{
\item{x}{(\code{vector})}
}
\description{
Returns number of unique values
}
\examples{
lunique(c(1, 1, 2, 2, 2, 3))
}
