% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapping_message}
\alias{mapping_message}
\title{Mapping Message}
\usage{
mapping_message(what, addtl = NULL)
}
\arguments{
\item{what}{(\code{character})\cr
Which values were mapped.}

\item{addtl}{(\code{character})\cr
An additional, specific message to append to the generic message.}
}
\value{
Prints a message to the console.
}
\description{
Message to display to inform the user that the function being used relies on
GEMINI SME mapped values.
}
