% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlaps.R
\name{min_result_value}
\alias{min_result_value}
\title{Calculates minima when input vector is not empty, else returns NA.}
\usage{
min_result_value(x)
}
\arguments{
\item{x}{(\code{numeric})\cr
A vecctor of numerical values.}
}
\value{
(\code{numeric})\cr
A numerical value or NA
}
\description{
Calculates minima when input vector is not empty, else returns NA.
}
\details{
This is a helper function to suppress default warning message from `base::min()`` function
when all elements in the input vector is NA, which can be problemetic for unit testing.
Default to remove NA values in minima calculation.
}
