% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_imaging.R
\name{n_imaging}
\alias{n_imaging}
\title{Compute the number of radiology tests per encounter}
\usage{
n_imaging(dbcon, cohort, exclude_ed = FALSE)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A \code{DBI} database connection to any GEMINI database.}

\item{cohort}{(\code{data.frame} or \code{data.table})
Cohort table with all relevant encounters of interest, where each row
corresponds to a single encounter. Must contain GEMINI Encounter ID
(\code{genc_id}).}

\item{exclude_ed}{(\code{logical})
Whether to exclude tests in emergency department. When set to \code{TRUE}, tests performed
in ED (testing time before admission time) will not be counted. When set to \code{FALSE},
tests will not be filtered by time and all tests in radiology table will be counted.
Please be aware that this may include tests before triage time, tests after discharge
time, and tests without testing time.

Tests in ED are defined as \code{ordered_date_time} earlier than \code{admission_date_time}.
When \code{ordered_date_time} is not available, \code{performed_date_time} is used instead.}
}
\value{
data.table with the same number of rows as input \code{cohort}, with additional
derived numeric fields labelled as "n_img_xray_derived",  "n_img_ct_derived",
"n_img_mri_derived",  "n_img_us_derived", "n_img_other_derived",
"n_img_int_derived" and "n_img_ct_mri_us_derived".
}
\description{
\code{n_imaging} returns the number of radiology tests for hospital admission.
}
\details{
This function takes a list of admissions and a GEMINI database connection to
generate numeric fields counting the number of different radiology tests for
each admission.

Imaging table in the database should include field that classifies each test
into below 7 categories.
\itemize{
\item 1: X-Ray
\item 2: CT
\item 3: MRI
\item 4: Ultrasound
\item 5: Other
\item 6: Interventional radiology procedures
\item 7: Echo (Excluded)
}

Therefore, this function should be run after some standardization efforts on
imaging table (performed by GEMINI team). Currently, the number of advanced
radiology tests is one of the performance metrics in \href{https://www.hqontario.ca/Quality-Improvement/Practice-Reports/MyPractice-General-Medicine}{MyPracticeReport}. CT,
MRI and Ultrasound are considered advance imaging tests, which can be
retrieved by the derived variable (\code{n_img_ct_mri_us_derived}).
}
\note{
Currently, the function does not take radiology data coverage into account. For
patients without imaging tests, the function will return 0 in result columns.
User should check radiology data coverage and decide whether the imputed
\code{0}s are appropriate or not.
}
\section{Warning}{

Function returns \code{data.table} with id field and several numeric fields. By
design, function will not return any NA values. When one tries to left-join
the output of this function with another table (another list of admissions in
the left), make sure list of admissions (or patient) aligns in both tables.
}

