% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_rbc_transfusions.R
\name{n_rbc_transfusions}
\alias{n_rbc_transfusions}
\title{Count the number Red Blood Cell (RBC) Transfusions per encounter}
\usage{
n_rbc_transfusions(dbcon, cohort, exclude_ed = FALSE)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A \code{DBI} database connection to any GEMINI database.}

\item{cohort}{(\code{data.frame} or \code{data.table})
Cohort table with all relevant encounters of interest where each row
corresponds to a single encounter. Must contain the following columns:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{hospital_num} (\code{integer}): Hospital number
}}

\item{exclude_ed}{(\code{logical})
Whether to exclude transfusions in emergency department. When set to \code{TRUE},
transfusions performed in ED (before admission time) will not be counted.
When set to \code{FALSE}, transfusions will not be filtered by time and all RBC
transfusions in \code{transfusion} table will be counted. Please be aware that this
may include transfusions before triage time, transfusions after discharge time,
and transfusions without issuing time.

Transfusions in ED are defined as \code{issue_date_time} earlier
than \code{admission_date_time}. Transfusions with missing \code{issue_date_time} will be
excluded when \code{exclude_ed} is set to \code{TRUE}.}
}
\value{
(\code{data.table})\cr
Table with three columns: \code{genc_id}, \code{n_app_rbc_transfusion_derived} (number
of appropriate RBC transfusions), and \code{n_rbc_transfusion_derived} (total of
all RBC transfusions). Encounters without any transfusion will get a 0.
}
\description{
This function calculates the number of appropriate transfusions and total of
all transfusions per encounter.
The definition of an appropriate transfusion can be found on the sample
\href{https://www.hqontario.ca/Portals/0/documents/qi/practice-reports/general-medicine-sample-report.html#transfusions-qi}{My Practice Report}.
A transfusion is defined by procedure in the below codes:
\enumerate{
\item \href{https://athena.ohdsi.org/search-terms/terms/4144461}{4144461}.
\item \href{https://athena.ohdsi.org/search-terms/terms/4137859}{4137859}
\item \href{https://athena.ohdsi.org/search-terms/terms/4022173}{4022173}
}

An appropriate transfusion also requires that the patients pre-transfusion
hemoglobin levels are  below 80 g/L. The Athena definition of a hemoglobin
test:
\href{https://athena.ohdsi.org/search-terms/terms/3000963}{3000963}

Red blood cell transfusions with no hemoglobin measurement within 48 hours
prior to the transfusion are excluded. These scenarios are rare, typically
occurring in approximately 2\% of blood transfusions in GEMINI data.
}
\note{
Transfusion data from two hospitals (105 and 106) with known historical
data quality issues is automatically excluded for encounters with a
\code{discharge_date_time} before 2021-01-01 00:00. Any \code{genc_ids} from those
sites meeting this criterion are not included in the returned output.
When merging the output of this function with another table, those
\code{genc_ids} should have a value of \code{NA}.

Currently, the function does not take \code{transfusion} or \code{lab} data coverage
into account. For patients without RBC transfusion, the function will return
\code{0} in result columns. User should check transfusion and lab data coverage and
decide whether the imputed \code{0}s are appropriate or not.
}
