% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_surgeries.R
\name{n_surgeries}
\alias{n_surgeries}
\title{Compute the number of surgical interventions per encounter}
\usage{
n_surgeries(cohort, interventions)
}
\arguments{
\item{cohort}{(\code{data.frame} or \code{data.table})
Cohort table with all relevant encounters of interest, where each row
corresponds to a single encounter. Must contain GEMINI Encounter ID
(\code{genc_id}).}

\item{interventions}{(\code{data.frame} or \code{data.table})
All intervention data of interest. This typically refers to the \code{ipintervention}
table containing all in-patient interventions
(see \href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}).
However, researchers may optionally choose to also include interventions performed
in the emergency department (\code{erintervention}) by merging the corresponding
tables outside of this function. Similarly, researchers can perform any other
relevant exclusions/inclusions (e.g., interventions within first 24h after admission)
prior to running this function.

This table must contain the \code{genc_id} and \code{intervention_code} (CCI code) fields in long format.
The intervention codes must be free from any punctuation or special characters.}
}
\value{
\code{data.table}
This function returns a table with all encounters identified by the \code{cohort} table input and
additional derived numeric field \code{n_surgeries} indicating the number of surgical interventions per \code{genc_id}.
If no surgical intervention was found, the corresponding \code{genc_id} will be returned with \code{n_surgeries = 0}.
}
\description{
\code{n_surgeries} returns the number of surgeries per \code{genc_id} based on CCI intervention codes.
}
\details{
This function takes a \code{cohort} and \code{intervention} data input and generates a
new numeric field corresponding to the number of surgical interventions per
\code{genc_id} in the cohort of interest.

A list of all CCI intervention codes that are classified as surgeries can be
found in the \href{https://www.cmaj.ca/content/cmaj/suppl/2020/11/10/192.46.E1440.DC1/200068-res-1-at.pdf}{supplemental material}
of this \href{https://www.cmaj.ca/content/192/46/E1440/tab-related-content}{paper}.
}
\section{Warning}{

By design, this function will not return any \code{NA} values. Any \code{genc_id} for which no surgery entries were found in the
intervention table are returned with \code{n_surgeries = 0}. When left-joining the output of this function with another table,
make sure list of genc_ids aligns in both tables.
}

\references{
Jerath, A., Sutherland, J., Austin, P. C., Ko, D. T., Wijeysundera, H. C., Fremes, S., ... & Wijeysundera, D. N.
(2020). Delayed discharge after major surgical procedures in Ontario, Canada: a population-based cohort study. CMAJ, 192(46), E1440-E1452.
}
