% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhd_ses_variables.R
\name{nbhd_ses_variables}
\alias{nbhd_ses_variables}
\title{Obtain commonly used neighbourhood-level socioeconomic (SES) variables}
\usage{
nbhd_ses_variables(dbcon, dissemination_area, cohort)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{dissemination_area\cr}{Version of dissemination_area chosen. Only 2016 or 2021 are valid inputs.}

\item{cohort\cr}{User specified cohort}
}
\value{
(\code{data.frame} | \code{data.table})\cr
This function returns a \code{data.table} merged with dissemination area ID and SES variables.
}
\description{
The \code{nbhd_ses_variables()} function merges SES variables based on dissemination area ID.
}
\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
                        dbname = "db",
                        host = "domain_name.ca",
                        port = 1234,
                        user = getPass("Enter user:"),
                        password = getPass("password"))

nbhd_ses_table <- nbhd_ses_variables(dbcon, 2016, cohort)
}

}
