% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_text}
\alias{normalize_text}
\title{Normalize string values}
\usage{
normalize_text(x, lemma = FALSE)
}
\arguments{
\item{x}{(\code{character})\cr
Text input to be normalized. Multiple values can be provided as a character vector.}

\item{lemma}{(\code{logical})\cr
If set to \code{TRUE}, singularizes x via \code{textstem::lemmatize_words()}.
For use in the RxNorm workflow: Note that lemmatization is an NLP technique that may not always convert drug
names or classification terms into their respective singular form, but this processing is still helpful to a
certain extent to help minimize variations due to plural forms.}
}
\description{
This function performs a series of text cleaning and normalization on text data.
For example, it is used in \code{prepare_pharm_for_validation()} to clean up RxNorm outputs for validation.
The operations include:
\itemize{
\item Converts text to ASCII encoding, such that special characters can be intepreted properly by R.
\item Converts all characters to lowercase.
\item Trims leading and trailing whitespace.
\item Removes leading and trailing periods.
\item Optional: Converts to the singular form through lemmatization.
}
}
\examples{
normalize_text(" Ámoxícíllíns. ")
}
