% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_theme.R
\docType{data}
\name{plot_colors}
\alias{plot_colors}
\title{GEMINI plotting colors}
\format{
An object of class \code{character} of length 3.
}
\usage{
plot_colors
}
\description{
GEMINI plotting colors
}
\keyword{datasets}
