% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_histograms}
\alias{plot_histograms}
\title{Plot histograms of multiple variables}
\usage{
plot_histograms(cohort, vars, binwidth)
}
\arguments{
\item{cohort}{(\code{data.frame} | \code{data.table})\cr Table containing cohort data.}

\item{vars}{(\code{character})\cr Character vector of variables to be plotted.}
}
\value{
(\code{ggplot})\cr A ggplot figure containing histograms of all variables
specified in \code{vars}.
}
\description{
This function ...
}
\examples{

\dontrun{


}

}
