% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hosp_time}
\alias{plot_hosp_time}
\title{Plot by hospital and time}
\usage{
plot_hosp_time(
  cohort,
  plot_var = NULL,
  time_var = "discharge_date_time",
  hosp_var = "hospital_num",
  hosp_group = NULL,
  facet_var = NULL,
  time_int = "month",
  func = "mean",
  plot_cat = NULL,
  show_overall = TRUE,
  line_width = 1,
  ylimits = NULL,
  min_n = 0,
  colors = gemini_colors(),
  return_data = FALSE
)
}
\arguments{
\item{cohort}{(\code{data.frame | data.table})\cr
A table containing the relevant cohort data to be plotted.}

\item{plot_var}{(\code{character})\cr
The name of the variable we wish to plot}

\item{time_var}{(\code{character})\cr
The name of the variable specifying time. Typically "discharge_date_time".}

\item{hosp_var}{(\code{character})\cr
The name of the variable specifying individual hospitals. By default:
\code{"hospital_num"} (unless \code{"hospital_num"} does not exist, in which case the
function will check for existance of \code{"hospital_id"} instead).}

\item{hosp_group}{(\code{character})\cr
The name of the variable used to group hospitals}

\item{facet_var}{(\code{character})\cr
The name of variable specifying facet plots}

\item{time_int}{(\code{character})\cr
Time interval to plot by (x-axis intervals). Currently, the function can
automatically calculate any of the following time intervals for any date-time
variables:
\itemize{
\item "month" (default)
\item "quarter"
\item "year"
\item "fisc_year" for hospital fiscal year starting in April
\item "season"
}

For any other custom time intervals (e.g., weeks), users can calculate the
desired time interval prior to running this function. As long as this custom
time variable exists in the \code{cohort} input, users can then specify this
variable as the \code{time_int} input. For example, users could derive a weekly
time variable called \code{"week"} in their \code{cohort} table, and
then specify \code{time_int = "week"}.
If \code{time_var} is equal to any of \code{c("month", "quarter", "year", "fisc_year")}
and there is a user-provided variable of that same name in the \code{cohort}
input, the function will default to the variable that exists in the user-
provided input.}

\item{func}{(\code{character})\cr
The summary function used to aggregate data by time & hospital. Has to be one
of the following:
\itemize{
\item "mean" (default)
\item "median"
\item "\%" or "prct"/"perc" (for categorical variables), users also need to specify
\code{plot_cat} to specify which category to plot the percentage of (see below)
\item "n" or "count" to plot the count of rows per hospital/time period
\item "missing" or "na"
}}

\item{plot_cat}{Required when \code{func = "\%"} and/or when plotting categorical/factor variables.
Users need to specify the level of any categorical variables to be plotted.
For example, to plot the percentage of female encounters, \code{plot_cat = "F"}.
Multiple categories can be provided as a vector, e.g., to plot all non-male
encounters: \code{plot_cat = c("F", "O")}. If no \code{plot_cat} is specified, the
function will sort the unique entries of the specified \code{plot_var} and plot
the percentage of rows matching the highest value (e.g., for logical
variables, \verb{\% TRUE} will be plotted by default).
\code{plot_cat} can also be used for numeric variables, e.g., \code{plot_cat = 0} to
plot encounters with Charlson Comorbidity Index = 0.}

\item{show_overall}{(\code{logical})\cr
Flag indicating whether to plot thick line representing overall value across
hospitals.}

\item{line_width}{(\code{numeric})\cr
Width of individual lines. Summary line will be 2 * line_width.}

\item{ylimits}{(\code{numeric})\cr
Numeric vector specifying limits for y-axis e.g. c(0, 100).}

\item{min_n}{(\code{numeric})\cr
Minimum number of data points required for each hospital * time point
combination. Data points with cell count < \code{min_n} will be suppressed.}

\item{colors}{(\code{character})\cr
Character vector specifying line color(s).}

\item{return_data}{(\code{logical})\cr
Flag indicating whether to return a list of 2 data.tables with aggregated
data (\link{1} by hospital and \link{2} overall). If \code{FALSE} (default), will return plot.}
}
\value{
By default, returns \code{ggplot} figure. If \code{return_data = TRUE}, returns
\code{data.tables} containing aggregated data.
}
\description{
Function plotting a variable of interest by hospital and over time.

Creates a ggplot of the cohort data across time for a hospital
}
\seealso{
\code{vignette("plotting_functions", package = "Rgemini")}
}
