% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot descriptive summary statistics of multiple variables}
\usage{
plot_summary(
  data,
  plot_vars = NULL,
  facet_group = NULL,
  show_stats = TRUE,
  prct = FALSE,
  base_size = NULL,
  color = "lightblue",
  ...
)
}
\arguments{
\item{data}{(\code{data.frame} | \code{data.table})\cr
Table containing data to be plotted.}

\item{plot_vars}{(\code{character} | \code{list})\cr
Character vector or list of variables to be plotted. If no \code{plot_vars} input
is provided, the function will automatically plot all variables, ignoring any
encounter/patient/physician IDs and date-time variables.}

\item{facet_group}{(\code{character})\cr
Name of variable to be used as facet variable. This only works if \code{plot_vars}
only specifies 1 variable to be plotted. Users can then create separate
subplots per \code{facet_group} level, for example, to plot separate histograms/
barplots for each hospital (\code{facet_group = "hospital_num"})/}

\item{show_stats}{(\code{logical})\cr
Flag indicating whether to show descriptive stats above each plot.}

\item{prct}{(\code{logical})\cr
Flag indicating whether y-axis labels should show percentage (\%). If \code{FALSE}
(default), counts (n) will be shown.}

\item{base_size}{(\code{numeric})\cr
Numeric input defining the base font size (in pts) for each subplot.
By default, the function will automatically determine an appropriate size
depending on the number of subplots (\code{base_size = 11} if a single subplot).}

\item{color}{(\code{character})\cr
Plotting color used for "fill". Default is R's built-in \code{"lightblue"}.}

\item{...}{\cr
Additional arguments passed to \code{ggpubr::ggarrange()} that allow for finer
control of subplot arrangement (e.g., \code{ncol}, \code{nrow}, \code{widths}, \code{heights},
\code{align} etc.; see \code{? ggarrange} for more details).}
}
\value{
(\code{ggplot})\cr A ggplot figure with subplots showing histograms/
barplots for all variables specified in \code{plot_vars}.
}
\description{
This function plots distributions (histograms/barplots) and shows basic
summary statistics (e.g., median [Q1, Q3], \% missing etc.) for multiple
variables.
}
\section{Note}{

These plots are not meant as publication-ready figures. Instead, the goal of
this function is to provide a quick and easy means to visually inspect the
data and obtain information about distributional properties of a wide range
of variables, requiring just a single line of code.
}

\section{Additional inputs when providing \code{plot_vars} as a list}{

When \code{plot_vars} are provided as a list, users can specify additional
characteristics for each individual variable, such as:
\itemize{
\item \code{class} (\code{character}): variable type, e.g., \code{"numeric"}, \code{"character"},
\code{"logical"} etc.
\item \code{sort} (\code{character}): for categorical variables, whether to sort bars in
ascending (\code{"^a"}) or descending (starting with \code{"^d"}) frequency
\item \code{binwidth}/\code{bins}/\code{breaks}: for numeric/integer variables, specifying the
histogram bins
\item \code{normal}: for numeric/integer variables, whether to assume normal
distribution (will show mean [SD] if \code{show_stats = TRUE})
}
}

\examples{
# simulate GEMINI data table
admdad <- gemSim::dummy_admdad(
  n = 10000,
  n_hospitals = 20,
  time_period = c(2015, 2022)
)

## Providing plot_vars as a character vector
plot_summary(
  data = admdad,
  plot_vars = c("age", "gender", "discharge_disposition")
)

## Providing plot_vars as a list input
plot_summary(
  admdad,
  plot_vars = list(
    `Discharge disposition` = list(
      plot_var = "discharge_disposition",
      class = "character",
      sort = "desc"
    ),
    `# Days in ALC` = list(
      plot_var = "number_of_alc_days",
      binwidth = 1,
      breaks = seq(0, 7, 1)
    )
  )
)

}
\seealso{
\code{vignette("plotting_data_exploration", package = "Rgemini")}
}
