% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{GEMINI plot theme}
\usage{
plot_theme(
  base_size = 12,
  base_family = "",
  show_grid = FALSE,
  aspect_ratio = 1,
  ...
)
}
\arguments{
\item{base_size}{(\code{numeric})\cr}

\item{base_family}{(\code{numeric})\cr}

\item{show_grid}{(\code{logical})\cr}

\item{aspect_ratio}{(\code{numeric})\cr
Aspect ratio of plot}

\item{...}{\cr
Additional arguments passed to \code{theme()}.}
}
\value{
ggplot with specified theme
}
\description{
Common theme that can be applied to any ggplot object
}
