% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_plot.R
\name{plot_time_hospital}
\alias{plot_time_hospital}
\title{Plotting Hospitals results by time}
\usage{
plot_time_hospital(
  cohort,
  plot_variable,
  time_var,
  hospital_var,
  hospital_group,
  title,
  yaxis.limits,
  plot_level,
  func = mean,
  show.individual.hospitals = TRUE,
  summary.line.width = 1.5
)
}
\arguments{
\item{cohort}{The source data frame}

\item{plot_variable}{The name of the variable we wish to plot by time}

\item{time_var}{The name of the variable specifying time}

\item{hospital_var}{The name of the variable specifying individual hospitals}

\item{hospital_group}{(optional) The name of the variable used to group hospitals}

\item{title}{(optional) Plot title}

\item{yaxis.limits}{(optional) Limits for the y-axis e.g. c(0, 100)}

\item{plot_level}{(optional) For categorical variables, specify the level \% to plot}

\item{func}{The summary function used for groups (default is mean)}

\item{show.individual.hospitals}{(default true) Show lines with the default hospitals}

\item{summary.line.width}{(default 1.5) width of the summary line}
}
\value{
ggplot figure
}
\description{
Creates a ggplot of the cohort data across time for a hospital
}
