% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Suppress errors/messages/warnings}
\usage{
quiet(func)
}
\arguments{
\item{func}{Function to be run quietly}
}
\description{
Run function without showing any errors/warnings/printed messages.

Note that certain messages (e.g., from RPostgreSQL) cannot be suppressed.
}
