% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbc_transfusions.R
\name{rbc_transfusions}
\alias{rbc_transfusions}
\title{Count the number Red Blood Cell Transfusions per encounter}
\usage{
rbc_transfusions(con, ipadmdad)
}
\arguments{
\item{con}{\cr
Database connection to any GEMINI database}

\item{ipadmdad}{(\code{data.table})\cr
\code{ipadmdad} table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.}
}
\value{
\code{data.table}
It includes three columns. It includes genc_id,
n_app_rbc_transfusion_derived, and n_rbc_transfusion_derived.
It will include all rows from the ipadmdad parameter. Any encounter without a
transfusion will get a 0.
}
\description{
This function calculates appropriate transfusions and any transfusions.
The definition of an appropriate transfusion can be found on the sample
\href{https://www.hqontario.ca/Portals/0/documents/qi/practice-reports/general-medicine-sample-report.html#transfusions-qi}{My Practice Report}.
A transfusion is defined by procedure in the below codes:
\enumerate{
\item \href{https://athena.ohdsi.org/search-terms/terms/4144461}{4144461}.
\item \href{https://athena.ohdsi.org/search-terms/terms/4137859}{4137859}
\item \href{https://athena.ohdsi.org/search-terms/terms/4022173}{4022173}
}

An appropriate transfusion also requires that the patients pre-transfusion hemoglobin levels are  below 80 g/L.
The Athena definition of a hemoglobin test:
\href{https://athena.ohdsi.org/search-terms/terms/3000963}{3000963}
}
