% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmission.R
\name{readmission}
\alias{readmission}
\title{Compute readmission and episodes of care.}
\usage{
readmission(
  dbcon,
  elective_admit = TRUE,
  death = TRUE,
  MAID = FALSE,
  palliative = FALSE,
  chemo = FALSE,
  mental = FALSE,
  obstetric = FALSE,
  signout = FALSE,
  return_readmit_enc = FALSE,
  restricted_cohort = NULL,
  readm_win = c(7, 30)
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{elective_admit}{(\code{logical})\cr
If \code{TRUE} (default), elective episodes of care are not considered to be true readmissions, and are therefore
removed from the numerator. Specifically, if the first encounter of the (n)th episode of care is elective
(\code{admit_category = 'L'}), the (n-1)th episode of care is not considered to be followed by a readmission
(\code{readmit(n-1) = FALSE}).}

\item{death}{(\code{logical})\cr
If \code{TRUE} (default), episodes of care ending in in-hospital death are removed from the denominator. Specifically,
if the (n)th episode of care ended in death, it is impossible for a patient to be readmitted, and therefore the
episode of care is not considered in the readmission calculation (\code{readmit(n) = NA}).}

\item{MAID}{(\code{logical})\cr
If \code{TRUE} (default): Episodes of care that involve MAID are not considered to be true readmissions, and are therefore
removed from the numerator. Specifically, if any encounter of the (n)th episode of care involves MAID, the (n-1)th
episode of care is not considered to be followed by a readmission (\code{readmit(n-1) = FALSE}).
Note that episodes of care involving MAID end in death, however, they are NOT removed from the denominator unless
the \code{death} flag is set to \code{TRUE} as well (recommended approach).}

\item{palliative}{(\code{logical})\cr
If \code{TRUE} (default): Episodes of care involving palliative care are excluded from both the numerator and denominator.
Specifically, episode of care with palliative care as a type-M diagnosis are not considered to be true readmissions
(\code{readmit(n-1) = FALSE}). Additionally, similarly to death, palliative episodes of care cannot be
followed by future readmission and are therefore excluded from the denominator (\code{readmit(n) = NA})}

\item{chemo}{(\code{logical})\cr
If \code{TRUE} (default): Episodes of care that involve chemotherapy are not considered to be true readmissions, and are
therefore removed from the numerator. Specifically, if any encounter of the (n)th episode of care involves
chemotherapy, the (n-1)th episode of care is not considered to be followed by a readmission
(\code{readmit(n-1) = FALSE}).}

\item{mental}{(\code{logical})\cr
If \code{TRUE} (default): Episodes of care for mental health are excluded from both the numerator and denominator.
Specifically, episodes of care that fall under CIHI's Major Clinical Category (MCC) 17 are not considered to be true
readmissions and are therefore removed from the numerator (\code{readmit(n-1) = FALSE}). Additionally, they
are not considered to be acute care and are therefore not included in the readmission denominator
(\code{readmit(n) = NA})}

\item{obstetric}{(\code{logical})\cr
If \code{TRUE} (default): Episodes of care that involve obstetric delivery are not considered to be true readmissions, and
are therefore removed from the numerator. Specifically, if any encounter of the (n)th episode of care involves
obstetric delivery, the (n-1)th episode of care is not considered to be followed by a readmission
(\code{readmit(n-1) = FALSE}).}

\item{signout}{(\code{logical})\cr
If \code{TRUE} (default), episodes of care where the last encounter was self sign-out/left against medical advice (LAMA)
are removed from the denominator. Specifically, if the (n)th episode of care ended in self-signout/LAMA, it is not
considered in the readmission calculation (\code{readmit(n) = NA}).}

\item{return_readmit_enc}{(\code{logical})\cr
If \code{TRUE}: The function will additionally return a column
for each readmission window the user provides, containing the \code{genc_id}
corresponding to the readmission encounter (i.e., encounter following the index encounter within n days).
This can be used to analyze characteristics of the readmission encounter (e.g., time of readmission, diagnoses/clinical outcomes at the time of readmission etc.).
For index encounters where the readmission flag is \code{FALSE}/\code{NA}, \verb{readmit(n)_genc_id} will be returned as \code{NA}.}

\item{restricted_cohort}{(\code{data.frame} or \code{data.table})\cr
User specified cohort that is a restricted subset of all encounters in DRM table "ipadmdad" (see
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}).
Must contain \code{genc_id} as the identifier. Default is \code{Null}, which loads the entire "ipadmdad"
table in the user-provided database (recommended approach).}

\item{readm_win}{(\code{integer} | \code{vector})\cr
readmission window(s) of interest, in days (by default: \code{readm_win = c(7,30)} to calculate 7- & 30-day readmission)}
}
\value{
This function returns a \code{data.table} containing \code{genc_id}, \code{AT_in_occurred}, \code{AT_out_occurred}, \code{epicare}
(see \code{\link[=episodes_of_care]{episodes_of_care()}}), \code{readmit7} (default), and \code{readmit30} (default). \code{readmit7} and \code{readmit30}
are logical variables indicating whether or not an episode of care was followed by a readmission. Note that
readmissions are attributed to the last encounter of each episode of care. Episodes of care where the last encounter
has \code{readmitX = NA} have been removed from the denominator.

Note: If the very last encounter of an episode of care has a transfer-out coding (but the encounter associated with
that transfer does not exist in the GEMINI database), the episode of care is excluded from the readmission
denominator (\code{readmit(n) = NA}). This is to avoid inaccurate attribution of a potential readmission to
the hospital associated with the intermediate (i.e., not the last) encounter of that episode of care.
}
\description{
\code{readmission.R} computes whether or not a patient associated with an episode of care was readmitted to a GEMINI
hospital within a time window of interest. The default readmission time window is 7 or 30 days.

In line with CIHI guidelines, readmission calculations are based on episodes of care instead of individual
encounters in order to avoid that linked transfers are counted as readmissions (also see documentation for
\code{episodes_of_care()}). Specifically, an episode of care refers to all contiguous inpatient hospitalizations to
any medical or intensive care service within the GEMINI network. Episodes involving inter-facility transfers are
linked regardless of diagnosis. An acute care transfer is assumed to have occurred if either of the following
criteria are met:
\itemize{
\item{An admission to a GEMINI hospital within 7 hours after discharge
from another GEMINI hospital, regardless of whether the transfer is coded}
\item{An admission to a GEMINI hospital within 7-12 hours after discharge
from another GEMINI hospital, and at least 1 hospital has coded the transfer: Coded transfers are based on the
DAD Institution From and Institution To fields}
}
For episodes of care involving acute care transfers, readmissions are attributed to the last hospital from which
the patient was discharged before readmission.
}
\details{
This function calls the \code{episodes_of_care()} function to link encounters with acute care transfers into episodes
of care. When conducting readmission calculations using the function output, the number of encounters where
\code{readmission == TRUE} is the numerator, and the number of eligible episodes of care (where \code{readmission != NA})
is the denominator. For example, 7-day readmission rate can be calculated by \code{mean(readmit7, na.rm=TRUE)}, where
\code{readmit7} is the variable returned by the \code{readmission()} function. This function is designed to be consistent
with \href{https://www.cihi.ca/en/indicators/all-patients-readmitted-to-hospital}{CIHI guidelines for readmission calculations}.
Please carefully review this reference before applying the function for research purposes. Users may choose to deviate from
the CIHI definition of readmission by setting any of the CIHI flags (see above) to \code{FALSE}.
}
\section{Warning}{

Warnings are produced if
\enumerate{
\item either the \code{death} or \code{elective_admit} flags are set to \code{FALSE},
\item the input contains a user-specified restricted cohort, or
\item more than 25\% of episodes of care are removed from the denominator due to the CIHI flags or buffer period.
}
}

\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
  dbname = "db",
  host = "domain_name.ca",
  port = 1234,
  user = getPass("Enter user:"),
  password = getPass("password")
)

## default readmission calculation (with elective = TRUE & death = TRUE by default)
readmission(dbcon)

## Readmission calculation following CIHI definition
readmission(dbcon,
  elective_admit = TRUE, death = TRUE, MAID = TRUE, palliative = TRUE,
  chemo = TRUE, mental = TRUE, obstetric = TRUE, signout = TRUE
)
}

}
\references{
\href{https://www.cihi.ca/en/indicators/all-patients-readmitted-to-hospital}{CIHI readmission guidelines}
}
\seealso{
\code{vignette("epicare_and_readmission", package = "Rgemini")}
}
