% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{render_cell_suppression.categorical}
\alias{render_cell_suppression.categorical}
\title{Render Cell Suppression (Categorical)}
\usage{
render_cell_suppression.categorical(x, ...)
}
\arguments{
\item{x}{(\code{character} or \code{factor})\cr
A categorical variable to summarize.}

\item{...}{\cr
Optionally accept a named \code{digits} (\code{integer}) or \code{single_level_binary} (\code{logical}) argument
which specifiesthe number of digits to round percentages to.}
}
\value{
named (\code{character})\cr
Concatenated with \code{""} to shift values down one row for proper alignment.
}
\description{
This is a custom render for \code{table1} categorical variables which performs GEMINI
"cell suppression" for any variable levels which contain fewer than 6 observations.

If the total number of these variable levels with fewer than 6 observations is less
than 6, \emph{all} cells for all levels of the variable must be censored (because it is
possible to indirectly deduce the missing counts otherwise).
}
\examples{
x <- factor(c(rep("a", times = nrow(mtcars)), "b"), levels = c("a", "b"))
render_cell_suppression.categorical(x)

x2 <- factor(c(rep("a", times = nrow(mtcars))), levels = c("a", "b"))
render_cell_suppression.categorical(x2)

y <- factor(
  c(rep("a", times = nrow(mtcars)), "b", "c", "d", "e", "f", "g"),
  levels = c("a", "b", "c", "d", "e", "f", "g")
)
render_cell_suppression.categorical(y)

z <- factor(
  c(
    rep("a", times = 100),
    rep("b", times = 50),
    rep("c", times = 7),
    rep("d", times = 2)
  ),
  levels = c("a", "b", "c", "d", "e")
)

render_cell_suppression.categorical(z)

}
