% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{render_cell_suppression.continuous}
\alias{render_cell_suppression.continuous}
\title{Render Cell Suppression (Continuous)}
\usage{
render_cell_suppression.continuous(x, ...)
}
\arguments{
\item{x}{(\code{numeric})\cr
A continuous variable to summarize.}
}
\value{
named (\code{character})\cr
Concatenated with \code{""} to shift values down one row for proper alignment.
}
\description{
This is a custom renderer for \code{table1} continuous variables which performs GEMINI
"cell suppression" for any variable levels which contain fewer than 6 observations.
}
\examples{
x <- 1:6
render_cell_suppression.continuous(x)

y <- 1:2
render_cell_suppression.continuous(y)

}
