% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{render_cell_suppression.default}
\alias{render_cell_suppression.default}
\title{Render Cell Suppression (Default)}
\usage{
render_cell_suppression.default(
  x,
  name,
  missing = any(is.na(x)),
  transpose = FALSE,
  render.empty = "NA",
  render.continuous = render_cell_suppression.continuous,
  render.categorical = render_cell_suppression.categorical,
  render.missing = render_cell_suppression.missing,
  ...
)
}
\arguments{
\item{x}{(\code{vector})\cr
A vector of numeric, factor, character or logical values.}

\item{name}{(\code{character})\cr
Name of the variable to be rendered (ignored).}

\item{missing}{(\code{logical})\cr
Should missing values be included?}

\item{transpose}{(\code{logical})\cr
Logical indicating whether on not the table is transposed.}

\item{render.empty}{(\code{character})\cr
A character to return when x is empty.}

\item{render.continuous}{(\code{function})\cr
A function to render continuous (i.e. numeric) values. Can also be a character string,
in which case it is passed to \code{table1:::parse.abbrev.render.code()}.}

\item{render.categorical}{(\code{function})\cr
A function to render categorical (i.e. factor, character or logical) values. Can also be a character string,
in which case it is passed to \code{table1:::parse.abbrev.render.code()}.}

\item{render.missing}{(\code{function})\cr
A function to render missing (i.e. NA) values. Can also be a character string,
in which case it is passed to \code{table1:::parse.abbrev.render.code()}. Set to \code{NULL} to ignore missing values.}

\item{...}{\cr
Further arguments, passed to \code{table1:::stats.apply.rounding()} or
\code{prettyNum()} for additional formatting (e.g., \code{big.mark = ","}).}
}
\value{
(\code{character})\cr
Summary of variable as a character vector with cell suppression applied.
}
\description{
This is a wrapper around the render functions for each type of variable.
}
