% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{render_cell_suppression.strat}
\alias{render_cell_suppression.strat}
\title{Render Cell Suppression (Strata)}
\usage{
render_cell_suppression.strat(label, n, transpose = F)
}
\arguments{
\item{x}{(\code{character} or \code{factor})\cr
A categorical stratification variable to summarize.}
}
\value{
named (\code{character})\cr
Concatenated with \code{""} to shift values down one row for proper alignment.
}
\description{
This is a custom render for \code{table1} stratification variables which performs GEMINI
"cell suppression" for any levels which contain fewer than 6 observations.

Note that even with strata variable cell suppression, it is possible to reverse-calculate
the total given the overall column. Therefore it is recommended to also hide the "Overall"
column in the call to \code{\link[table1:table1]{table1::table1()}}.
}
