% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{render_default.discrete}
\alias{render_default.discrete}
\title{Render Default (Discrete)}
\usage{
render_default.discrete(x, ...)
}
\arguments{
\item{x}{(\code{character} or \code{factor})\cr
A discrete variable to summarize.}

\item{...}{\cr
Further arguments, passed to \code{prettyNum()} for additional formatting (e.g.,
\code{big.mark = ","}).}
}
\value{
named (\code{character})\cr
Concatenated with \code{""} to shift values down one row for proper alignment.
}
\description{
This is the default render for discrete variables in the \code{table1} package. It will
generate a sum for each level of the variable.
}
\examples{
render_default.discrete(mtcars$vs)

}
