% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{render_median.continuous}
\alias{render_median.continuous}
\title{Render Median (Continuous)}
\usage{
render_median.continuous(x, ...)
}
\arguments{
\item{x}{(\code{character} or \code{factor})\cr
A continuous variable to summarize.}

\item{...}{\cr
Further arguments, passed to \code{table1:::stats.apply.rounding()}.}
}
\value{
named (\code{character})\cr
Concatenated with \code{""} to shift values down one row for proper alignment.
}
\description{
This is the default renderer for continuous variables in the \code{table1} package. It will
generate a formatted median with first and third quartiles for each level of the variable.
}
\examples{
render_median.continuous(mtcars$disp)

}
