% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{render_strict_cell_suppression.categorical}
\alias{render_strict_cell_suppression.categorical}
\title{Render Strict Cell Suppression (Categorical)}
\usage{
render_strict_cell_suppression.categorical(x)
}
\arguments{
\item{x}{(\code{character} or \code{factor})\cr
A categorical variable to summarize.}
}
\value{
named (\code{character})\cr
Concatenated with \code{""} to shift values down one row for proper alignment.
}
\description{
Strictly suppress any counts that are less than 6 with message. This differs
from the more conservative logic of \code{\link[=render_cell_suppression.categorical]{render_cell_suppression.categorical()}}.
}
\examples{
x <- factor(c(rep("a", times = nrow(mtcars)), "b"), levels = c("a", "b"))
render_strict_cell_suppression.categorical(x)

x2 <- factor(c(rep("a", times = nrow(mtcars))), levels = c("a", "b"))
render_strict_cell_suppression.categorical(x2)

y <- factor(
  c(rep("a", times = nrow(mtcars)), "b", "c", "d", "e", "f", "g"),
  levels = c("a", "b", "c", "d", "e", "f", "g")
)
render_strict_cell_suppression.categorical(y)

z <- factor(
  c(
    rep("a", times = 100),
    rep("b", times = 50),
    rep("c", times = 7),
    rep("d", times = 2)
  ),
  levels = c("a", "b", "c", "d", "e")
)

render_strict_cell_suppression.categorical(z)

}
