% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_schema.R
\name{response_schema}
\alias{response_schema}
\title{Generate a JSON Schema from a Response Template}
\usage{
response_schema(example)
}
\arguments{
\item{example}{A response template created by \code{obj_resp} or \code{arr_resp}.}
}
\value{
A list representing the inferred JSON Schema for the provided response template.
}
\description{
This function recursively generates a JSON Schema based on a response template.
The template must be created using either \code{obj_resp} (for a single object response)
or \code{arr_resp} (for an array response). The function infers the JSON Schema
by inspecting the data types of the provided fields without including the example values.
}
