% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_hospital_field}
\alias{return_hospital_field}
\title{Return Hospital Field}
\usage{
return_hospital_field(db)
}
\arguments{
\item{db}{(\code{DBIConnection})\cr
RPostgres DB connection.}
}
\value{
(\code{character})\cr
\code{hospital_id} or \code{hospital_num}, with preference given to \code{hospital_id} if it exists.
}
\description{
To accommodate differences in column names between databases, find the name of the column corresponding to
the hospital for downstream queries.
}
