% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rlnorm_trunc}
\alias{rlnorm_trunc}
\title{Sample a truncated log normal distribution}
\usage{
rlnorm_trunc(n, meanlog, sdlog, min, max, seed = NULL)
}
\arguments{
\item{n}{(\code{integer}) The length of the output vector}

\item{meanlog}{(\code{numeric}) The mean of the log normal distribution}

\item{sdlog}{(\code{numeric}) The standard deviation of the log normal distribution}

\item{min}{(\code{numeric}) The minimum value to truncate the data to.}

\item{max}{(\code{numeric}) The maximum value to truncate the data to.}

\item{seed}{(\code{integer}) Optional, a number for setting the seed for reproducible results}
}
\value{
A numeric vector following the log normal distribution, truncated to the specified range.
}
\description{
Sample from a log normal distribution using the \code{rlnorm} function
Truncate it to specified minimum and maximum values
}
