% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{rnorm_trunc}
\alias{rnorm_trunc}
\title{Sample a truncated normal distribution}
\usage{
rnorm_trunc(n, mean, sd, min_n, max_n, seed = NULL)
}
\arguments{
\item{n}{(\code{integer}) The length of the output vector}

\item{mean}{(\code{numeric}) The mean of the normal distribution}

\item{sd}{(\code{numeric}) The standard deviation of the normal distribution}

\item{min_n}{(\code{numeric}) The minimum value to truncate the data to.}

\item{max_n}{(\code{numeric}) The maximum value to truncate the data to.}

\item{seed}{(\code{integer}) Optional, a number for setting the seed for reproducible results}
}
\value{
A numeric vector following the normal distribution, truncated to the specified range.
}
\description{
Sample from a normal distribution using the \code{rnorm} function but truncate it to specified minimum and maximum values
}
