% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsn_trunc}
\alias{rsn_trunc}
\title{Sample a truncated skewed normal distribution}
\usage{
rsn_trunc(n, xi, omega, alpha, min, max, seed = NULL)
}
\arguments{
\item{n}{(\code{integer}) The length of the output vector}

\item{xi}{(\code{numeric}) The center of the skewed normal distribution}

\item{omega}{(\code{numeric}) The spread of the skewed normal distribution}

\item{alpha}{(\code{numeric}) The skewness of the skewed normal distribution}

\item{min}{(\code{numeric}) The minimum value to truncate the data to.}

\item{max}{(\code{numeric}) The maximum value to truncate the data to.}

\item{seed}{(\code{integer}) Optional, a number for setting the seed for reproducible results}
}
\value{
A numeric vector following the skewed normal distribution, truncated to the specified range.
}
\description{
Sample from a skewed normal distribution using the \code{rsn} function
Truncate it to specified minimum and maximum values
}
