% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{sample_icd}
\alias{sample_icd}
\title{Simulate ICD-10 Diagnosis Codes}
\usage{
sample_icd(n = 1, source = "comorbidity", dbcon = NULL, pattern = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated in Rgemini version 3.0.0.
}
