% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{sample_scu_date_time}
\alias{sample_scu_date_time}
\title{Sample SCU admission and discharge date times by genc_id}
\usage{
sample_scu_date_time(
  scu_cohort,
  use_ip_dates = TRUE,
  start_date = NULL,
  end_date = NULL,
  seed = NULL
)
}
\arguments{
\item{scu_cohort}{(\code{data.table}) The dummy data table requiring the addition of SCU admission and discharge
date time columns.}

\item{use_ip_dates}{(\code{logical}) Optional, whether the table \code{df1} contains information about inpatient admission and
discharge date times. If TRUE, the function will sample SCU data based on these date times and if not,
it will sample at random.}

\item{start_date}{(\code{Date}) Optional, the earliest date in the range for the SCU admissions in the dummy data table.
It is not used if \code{use_ip_dates} is TRUE.}

\item{end_date}{(\code{Date}) Optional, the latest date in the range for the SCU admissions in the dummy data table.
It is not used if \code{use_ip_dates} is TRUE.}

\item{seed}{(\code{integer}) Optional, an integer for setting the seed for reproducible results.}
}
\value{
(\code{data.table}) A copy of the input, \code{scu_cohort}, will be returned.
It will contain the same fields, with the addition of:
\itemize{
\item \code{scu_admit_date_time} (\code{character}): the date and time of admission to the SCU
\item \code{scu_discharge_date_time} (\code{character}): the date and time of discharge from the SCU
}
}
\description{
The "ipscu" data table is a long format data table with multiple repeats for each genc_id.
They must be sampled such that SCU stays occur between inpatient admission and discharge dates and times,
and that if a patient has multiple SCU stays, they do not overlap.
It is a helper function for \code{dummy_ipscu}.
}
\examples{
cohort <- dummy_ipadmdad()
sample_scu_date_time(cohort, use_ip_dates = TRUE, seed = 1)
sample_scu_date_time(cohort,
  use_ip_dates = FALSE,
  start_date = as.Date("2015-01-01"), end_date = as.Date("2023-12-31")
)

}
