% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_time_shifted}
\alias{sample_time_shifted}
\title{Chopped, skewed normal distribution for time variables}
\usage{
sample_time_shifted(nrow, xi, omega, alpha, min = 0, max = 48, seed = NULL)
}
\arguments{
\item{nrow}{(\code{integer}) The number of data points to sample}

\item{xi}{(\code{numeric}) The center of the skewed normal distribution}

\item{omega}{(\code{numeric}) The spread of the skewed normal distribution}

\item{alpha}{(\code{numeric}) The skewness of the skewed normal distribution}

\item{min}{(\code{numeric}) Optional, a minimum value to left truncate the value to; the default value is set to 0.}

\item{max}{(\code{numeric}) Optional, a maximum value to right truncate the value to; the default value is set to 48.}

\item{seed}{(\code{integer}) Optional, an integer for setting the seed for reproducible results.}
}
\value{
A numeric vector following the specified distribution.
}
\description{
The function samples from a skewed normal distribution using \code{rsn} to obtain time of day data in hours. Values greater than 24 are subtracted by 24 (moved to the next day) so that a real time variable is observed.
}
