% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{scale_color_gemini}
\alias{scale_color_gemini}
\title{Wrapper function for \code{ggplot2::scale_color_manual()} that applies GEMINI
color palettes to ggplot objects.}
\usage{
scale_color_gemini(palette = 1)
}
\arguments{
\item{palette}{(\code{character} or \code{numeric})\cr
Name or index of color palette. Run \code{plot_color_palettes()} to see options.
By default, the first palette ("GEMINI Rainbow") is returned.}
}
\description{
Wrapper function for \code{ggplot2::scale_color_manual()} that applies GEMINI
color palettes to ggplot objects.
}
