% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_holidays_ON.R
\name{stat_holidays_ON}
\alias{stat_holidays_ON}
\title{Flag statuatory holidays in Ontario}
\usage{
stat_holidays_ON(data, date_column, include_observed_holidays = T)
}
\arguments{
\item{data}{data frame containing date column in format YYYY-MM-DD}

\item{date_column}{date column name to merge with holiday table}

\item{include_observed_holidays}{includes observed holidays in holiday table when TRUE}
}
\value{
This function returns a given data frame with additional columns as fields
holiday_day: the day of the week when the holiday occurred
holiday_name: the name of the holiday occuring on the given date
}
\description{
This function merges a given data frame with a holiday table to flag which dates are holidays/observed holidays
}
