% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_hospitalizations.R
\name{birth_hospitalizations}
\alias{birth_hospitalizations}
\title{Birth Hospitalizations flag}
\usage{
birth_hospitalizations(
  dbcon,
  cohort,
  include_er = FALSE,
  healthy_birth = FALSE
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{cohort}{(\code{data.table} or \code{data.frame})\cr
Table with all relevant encounters of interest, where each row corresponds
to a single encounter. Must contain \code{genc_id}.}

\item{include_er}{(\code{logical})\cr
Whether to include diagnosis codes from the \code{erdiagnosis} table in addition
to \code{ipdiagnosis} when identifying \code{"Z38"} codes.}

\item{healthy_birth}{(\code{logical})\cr
Placeholder argument, currently unused. Reserved for a future
\code{healthy_birth} flag}
}
\value{
(\code{data.table})\cr
For each unique \code{genc_id} in the input \code{cohort}, returns the \code{genc_id} and
a derived logical column \code{birth_hospitalization}:
\itemize{
\item \code{TRUE}: at least one \code{"Z38"} diagnosis code found and
\code{paeds_age_weeks = 0}.
\item \code{FALSE}: at least one diagnosis code found, but the encounter does not
meet both criteria above (no \code{"Z38"} code, or \code{paeds_age_weeks} not
equal to 0).
\item \code{NA}: no diagnosis codes found for the encounter in the table(s) queried,
or \code{paeds_age_weeks} is missing (flag cannot be determined).
}
}
\description{
Flag encounters corresponding to a birth hospitalization.
An encounter is flagged \code{TRUE} if it has at least one diagnosis code
containing \code{"Z38"} in the \code{ipdiagnosis} table (or, if
\code{include_er = TRUE}, both \code{ipdiagnosis} and
\code{erdiagnosis} tables), and a paediatric age in
weeks of 0 (\code{paeds_age_weeks = 0}) in the \code{admdad} table.
}
\examples{
\dontrun{
# establish a DB connection
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
  dbname = "db",
  host = "domain_name.ca",
  port = 1234,
  user = "user",
  password = getPass("Enter password:")
)

# flag birth hospitalizations for a cohort of encounters
birth_flags <- birth_hospitalizations(
  dbcon = dbcon, cohort = cohort,
  include_er = TRUE
)
}

}
