% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_product_lookup.R
\docType{data}
\name{blood_product_lookup}
\alias{blood_product_lookup}
\title{Blood product lookup table}
\format{
A data frame with 23 rows and 3 columns:
\describe{
\item{blood_product_mapped_omop}{Mapped OMOP codes for each blood product}
\item{blood_product_mapped_raw}{Raw names for each blood product}
\item{prob}{Relative proportions of each product, 0-1}
}
}
\usage{
blood_product_lookup
}
\description{
A dataset containing common blood product codes, their raw names, and relative proportions.
}
\keyword{datasets}
