% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_creation.R
\name{cohort_creation}
\alias{cohort_creation}
\title{Cohort creation}
\usage{
cohort_creation(
  cohort,
  labels,
  exclusion_flag = NULL,
  show_prct = TRUE,
  group_var = NULL,
  ...
)
}
\arguments{
\item{cohort}{(\code{list})
A list of cohort tables at each step of the cohort inclusion/exclusion steps.
The function will count the number of rows in each list item to determine the
cohort size at each step.}

\item{labels}{(\code{character})
Vector containing a description for each inclusion/exclusion step (needs to
be in the same order as corresponding list items in \code{cohort} input).}

\item{exclusion_flag}{(\code{logical})
A vector indicating whether a given cohort creation step should be
shown as an exclusion. If \code{TRUE} the number and percentage of rows that
were removed (rather than included) will be shown.

By default, all cohort steps will be interpreted as inclusion steps.}

\item{show_prct}{(\code{logical})
Flag indicating whether to show percentage values. If \code{FALSE}, only raw
counts will be shown.}

\item{group_var}{(\code{character})
Name of a grouping variable. If provided, cohort numbers will be stratified
by each level of the grouping variable (in addition to overall cohort
numbers).}

\item{...}{Additional parameters that will be passed to \code{prettyNum} for additional
formatting of numbers (e.g., \code{big.mark = ","}).}
}
\value{
A data.table showing the cohort numbers as a count (and percentage) at each
step of the cohort creation.
}
\description{
This function creates a table showing the cohort size (e.g., number of
encounters) that were included/excluded at each step of the cohort creation.
}
\examples{
my_data <- Rgemini::dummy_ipadmdad(10000, n_hospitals = 5) \%>\% data.table()

cohort_table <- cohort_creation(
  cohort = list(
    my_data,
    my_data[gender == "F"],
    my_data[gender == "F" & age >= 65],
    my_data[gender == "F" & age >= 65 & !grepl("^7", discharge_disposition)]
  ),
  labels = c(
    "All GEMINI encounters",
    "Gender = Female",
    "Age >= 80",
    "In-hospital death"
  ),
  exclusion_flag = c(FALSE, FALSE, FALSE, TRUE),
  group_var = "hospital_num"
)

}
