% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_sets}
\alias{compare_sets}
\title{Compare two sets to get the number of unique and common elements in each set}
\usage{
compare_sets(x, y, dates = FALSE, orders = c("ymd", "ymd HM", "ymd HMS"))
}
\arguments{
\item{x}{(\code{vector}) \cr
The first set to be compared}

\item{y}{(\code{vector}) \cr
The second set to be compared}

\item{dates}{(\code{logical})\cr
If set to TRUE, will attempt to convert x & y into dates, and then compare. Set to FALSE by default.}

\item{orders}{(\code{vector})\cr
Allows user to specify potential date or date-time formats they would like
to compare. The function will check for \code{"ymd"}, \code{"ymd HM"}, and \code{"ymd HMS"}
by default.

For example, users can set \code{orders = "dmy"} to compare vectors in \code{dmy}
format.}
}
\value{
(\code{data.table})\cr
A table showing the number of elements in both vectors, in the first vector only, and in the second vector only
}
\description{
This function takes in two vectors and returns the number
of overlapping elements, along with the number of unique elements.
The function can compare sets of characters, numerics, and dates.
}
\examples{
compare_sets(c(1:10), c(5:10), dates = FALSE)
}
