% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_datetime}
\alias{convert_datetime}
\title{Convert date-time variables into POSIXct/POSIXt format.}
\usage{
convert_datetime(
  dt_var,
  format = "ymd HM",
  addtl_missing = "",
  addtl_invalid = "",
  addtl_missing_ts = ""
)
}
\arguments{
\item{dt_var}{\cr
A vector containing the date-time values to be converted to the specified
format. The vector can be of any class (e.g., "character" or "POSIXct"). If
users have already pre-processed date-time variables into the expected
format (e.g., using \code{lubridate::ymd_hm()}) prior to running the function,
the function will simply return the original \code{dt_var}.}

\item{format}{(\code{character})\cr
Format to be used to parse \code{dt_var}, such as \code{"ymd HM"} (default),
\code{"ymd HMS"}, \code{"ymd"} etc. Will be passed to \code{lubridate::parse_date_time()}.

Multiple acceptable formats can be specified by providing a character vector,
e.g. \code{format = c("ymd HM", "ymd HMS", "ymd")}.

For more details, see \code{orders} input argument in
\link[lubridate]{parse_date_time}.}

\item{addtl_missing}{(\code{character})\cr
Additional message to be added to warning about missing entries (if any).}

\item{addtl_invalid}{(\code{character})\cr
Additional message to be added to warning about invalid entries (if any),
i.e., non-missing entries that could not be parsed into specified format.
For example, \code{"2020-01-01"} is invalid if \code{format = "ymd HM"}.}

\item{addtl_missing_ts}{(\code{character})\cr
If required \code{format} contains timestamp (e.g., \code{"ymd HM"} or \code{"ymd HMS"}):
Additional message to be added to warning about entries with missing
timestamp (i.e., date-only entries).}
}
\value{
(\code{POSIXct} | \code{POSIXt})\cr
Returns converted \code{dt_var} parsed according to specified date-time \code{format}.

Will also return the following warning messages (if applicable):
\enumerate{
\item Number (\%) of entries in \code{dt_var} that are missing (NA, "", " ")
\item Number (\%) of entries in \code{dt_var} that could not be parsed into specified
format
\item For formats containing timestamps: Number (\%) of entries in \code{dt_var} that
only contain date information
}

Any missing/invalid entries will be returned as \code{NA}.
}
\description{
This function converts date-time variables into a user-specified format.
\code{convert_date_time} is called by other \code{Rgemini} functions to make sure that
any date-time inputs are in the expected format (typically \code{"ymd HM"}). It
can also be used independently of other \code{Rgemini} functions to clean up
date-times for analyses.
The function additionally returns warning messages about missing/invalid
date-time entries that can be useful for debugging.
}
\examples{
my_date_time <- c("2020-01-01 12:00", "2021-03-06 09:25")
convert_datetime(my_date_time, format = "ymd HM")
}
