% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coverage.R
\name{data_coverage}
\alias{data_coverage}
\title{Check data coverage}
\usage{
data_coverage(
  dbcon,
  cohort,
  table,
  plot_timeline = TRUE,
  plot_coverage = TRUE,
  ...
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{cohort}{(\code{data.frame} or \code{data.table})
Cohort table with all relevant encounters of interest, where each row
corresponds to a single encounter. Must contain the following columns:
\itemize{
\item \code{genc_id}: GEMINI Encounter ID
\item \code{hospital_num} | \code{hosital_id}: Hospital identifier
\item \code{discharge_date_time}
}}

\item{table}{(\code{character})
Which table(s) to include. If multiple, specify a character vector
(e.g., \code{table = c("lab", "pharmacy", "radiology")}).}

\item{plot_timeline}{(\code{logical})
Flag indicating whether to plot an overview of data availability timelines
by site and table.
\strong{Note:} This plot does not show actual coverage (e.g., \% encounters with
available data), but instead, only shows a rough overview of min-max dates,
and illustrates major gaps in data availability. Specifically, gaps in this
plot illustrate time periods where no data were available for at least 30
consecutive days. Importantly, for time periods without any gaps, data may
only be partially available (i.e., data coverage might be low or certain
columns/values of interest may be missing). Users should further inspect the
coverage plots (\code{plot_coverage = TRUE}) and perform additional customized
checks based on their needs.}

\item{plot_coverage}{(\code{logical})
Flag indicating whether to plot data coverage.
These plots show the percentage of \code{genc_ids} with an entry in the table(s)
of interest. Data coverage is plotted by hospital and discharge month
(because GEMINI data are pulled based on encounters' discharge date).
Users should carefully inspect these plots for any drops/gaps in coverage
that may not be visible in the timeline plots (see \code{plot_timeline} above).}

\item{...}{Additional inputs that can be passed to \code{plot_over_time}, e.g.:
\itemize{
\item \code{time_int = "fisc_year"} (instead of default \code{"month"})
}}
}
\value{
data.table object with the same number of rows as input \code{cohort}, with
additional derived column(s) containing a flag indicating whether \code{genc_id}
falls within data coverage period for the table of interest.
}
\description{
This function returns a flag for each \code{genc_id} indicating whether data for
a table of interest (e.g., "lab") were in principle available during the time
period in which the \code{genc_id} was discharged.

By default, the function will also return several plots and additional
tables that show an overview of data availability timelines by hospital and
table, and data coverage (\% \code{genc_ids} with an entry in a given table for
each month).
}
\section{Warning}{

}

