% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coverage.R
\name{data_coverage}
\alias{data_coverage}
\title{Check data coverage}
\usage{
data_coverage(
  dbcon,
  cohort,
  table,
  plot_timeline = TRUE,
  plot_coverage = TRUE,
  ...
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{cohort}{(\code{data.frame} or \code{data.table})
Cohort table with all relevant encounters of interest, where each row
corresponds to a single encounter. Must contain the following columns:
\itemize{
\item \code{genc_id}: GEMINI Encounter ID
\item \code{hospital_num} | \code{hosital_id}: Hospital identifier
\item \code{discharge_date_time}
}}

\item{table}{(\code{character})
Which table(s) to include. If multiple, specify a character vector
(e.g., \code{table = c("lab", "pharmacy", "radiology")}).}

\item{plot_timeline}{(\code{logical})
Flag indicating whether to plot minimum - maximum available dates per site
per table.}

\item{plot_coverage}{(\code{logical})
Flag indicating whether to plot coverage (\% \code{genc_ids} with entry in table).}

\item{...}{Additional inputs that can be passed to \code{plot_over_time}, e.g.:
\itemize{
\item \code{time_int = "fisc_year"} (instead of default \code{"month"})
}}
}
\value{
data.table object with the same number of rows as input "cohort", with
additional derived column(s) containing a flag indicating whether \code{genc_id}
falls within data-coverage period for the table of interest.
}
\description{
This function returns a flag for each \code{genc_id} indicating whether data for
a table of interest (e.g., "lab") were in principle available for that
\code{genc_id} based on a data coverage table in the database.
}
\section{Warning}{

}

