% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coverage.R
\name{data_coverage}
\alias{data_coverage}
\title{Check data coverage}
\usage{
data_coverage(
  dbcon,
  cohort,
  table,
  plot_timeline = TRUE,
  plot_coverage = TRUE,
  hospital_label = NULL,
  as_plotly = FALSE,
  custom_dates = NULL,
  ...
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{cohort}{(\code{data.frame} or \code{data.table})
Cohort table with all relevant encounters of interest, where each row
corresponds to a single encounter. Must contain the following columns:
\itemize{
\item \code{genc_id}: GEMINI Encounter ID
\item \code{hospital_num} | \code{hospital_id}: Hospital identifier
\item \code{discharge_date_time}
}}

\item{table}{(\code{character})
Which table(s) to include. If multiple, specify a character vector
(e.g., \code{table = c("lab", "pharmacy", "radiology")}).

For HPC4Health users: Please specify the full table name as it is listed in
your datacut (e.g., \code{"admdad_subset"} instead of only \code{"admdad"}).}

\item{plot_timeline}{(\code{logical})
Flag indicating whether to plot an overview of data timelines by hospital and
table.
\strong{Note:} This plot only shows a rough overview of min-max dates per table
for each hospital. Gaps in this plot illustrate time periods where no data
were available for at least 30 consecutive days. Importantly, for time
periods without any gaps, overall data volume may still be low and certain
entries/columns may still be missing. Users should further inspect the
coverage plots (\code{plot_coverage = TRUE}) and perform additional customized
checks based on their needs.}

\item{plot_coverage}{(\code{logical})
Flag indicating whether to plot data coverage. These plots show the
percentage of \code{genc_ids} with an entry in the table(s) of interest. Data
coverage is plotted by hospital and discharge month (because GEMINI data are
pulled based on encounters' discharge date). Users should carefully inspect
these plots for any drops/gaps in coverage that may not be visible in the
timeline plots (see \code{plot_timeline} above). When plotting the \code{admdad} table,
the function will plot the number of encounters by discharge month to show
overall data volume (note: by definition, 100\% of \code{genc_ids} in GEMINI data
have an entry in \code{admdad}).}

\item{hospital_label}{(\code{character})
Optional: Name of variable in \code{cohort} table that corresponds to custom
label for hospitals (e.g., letters A-E instead of hospital_num 101-105).
Will be used for plotting purposes.}

\item{as_plotly}{(\code{logical})
Will return any figures as interactive plots using \code{plotly}. Note that this
may affect plot aesthetics.
The flag will be ignored if the \code{plotly} package is not installed.}

\item{custom_dates}{(\code{data.frame}|\code{data.table})
Optional input allowing users to specify a customized timeline to be
included for a given hospital*table combination. The user-provided input
overwrites the corresponding row(s) in the \code{lookup_data_coverage} table.
This can be used to exclude time periods (e.g., due to data quality
issues) and generate customized timeline plots.
For example, let's say you identified a data quality issue in the
transfusion table at hospital 104 for discharge dates < 2019-01-01.
To only include transfusion data for encounters discharged after this
time period, specify:
\code{custom_dates <- data.frame( data = "transfusion", hospital_num = 104, min_date = "2019-01-01", max_date = "2022-06-30" )}

This will overwrite the \code{min_date}/\code{max_date} values for site 104
transfusion data in the \code{lookup_data_coverage} table (data for all
other hospital*table combinations will remain the same). Additionally, the
coverage flags by \code{genc_id} (in returned \code{coverage_flag_enc}) and timeline plot
(see \code{plot_timeline}) will be adjusted according to the user-provided dates.
The coverage plot (see \code{plot_coverage}) is not affected by the user-specified
the \code{custom_dates} input.}

\item{...}{Additional inputs that can be passed to control plot aesthetics in
\code{plot_timeline} or \code{plot_coverage} plots, such as:
\itemize{
\item \code{base_size}: Font size (default = 12)
\item \code{colors}: Plot color(s) (default = gemini_colors(1))
\item \code{color_group}: Name of variable in cohort specifying color grouping of
hospitals (e.g., Teaching/Non-teaching); for the timeline plot, this is
only applied when plotting a single table (otherwise, color grouping is
applied to different table names by default)
}

For coverage plots only (inputs are passed to \code{plot_over_time()}):
\itemize{
\item \code{time_int}: Time interval used to aggregate data (e.g., by \code{"month"}
\link{default}, \code{"quarter"}, or \code{"year"})
\item \code{ylimits}: Limits of y-axis (e.g., \code{c(0, 50)})
\item \code{scales}: Passed to facet wrap to control if y-scales are \code{"fixed"} (default)
or \code{"free"} (only works if no \code{ylimits} specified)
}}
}
\value{
If the plotting flags are set to \code{FALSE}, this function will return a single
\code{data.table} object with a flag for each \code{genc_id} indicating whether the
encounter was discharged during a time period in which data for a given
table (e.g., \code{"lab"}) were \emph{in principle} available. If the flag is \code{FALSE},
the \code{genc_id} was dicharged during a time period where GEMINI did not receive
any data for the table of interest. If the flag is \code{TRUE}, the \code{genc_id} was
discharged during a time period where GEMINI received \emph{some} data from a
given hospital (however, coverage may still be low, so users are advised to
perform additional coverage checks, e.g., by using the plotting features of
this function).

When the plotting flags are set to \code{TRUE} (default), the function will
return additional data tables (\code{output[["data"]]}) and plots
(\code{output[["plots"]]}):
\itemize{
\item If \code{plot_timeline = TRUE}: Returns \code{timeline_plot} showing an overview of
data timelines by hospital for each table. Will also return the corresponding
data table \code{timeline_data} that lists the min - max dates of each table per
site (in long format, where tables with multiple rows per hospital indicate
gaps in the data timeline)
\item If \code{plot_coverage = TRUE}: Returns \code{coverage_plot} showing the percentage
of \code{genc_ids} with an entry in a given table by discharge month and hospital.
Will also return the plotted data as \code{coverage_data} to facilitate further
inspection.
}
}
\description{
This function facilitates data coverage checks that analysts should perform
during data exploration & cohort generation. Specifically, this function
checks for "gaps" in the data timeline (e.g., are there any hospitals/time
periods for which GEMINI did not receive any data at all?), and also provides
more detailed insights into data volume & coverage by month and hospital
(e.g., what percentage of encounters have an entry in a given table?).
All checks performed by this function are applied at the table level (i.e.,
missing values in individual columns are not considered and should be checked
separately).
}
\details{
\code{data_coverage} provides analysts with a tool to inform their decisions
about which hospitals/time periods to include in their analyses,
depending on the data tables of interest. For example, if a project relies on
lab data (e.g., \code{mlaps} variable), users should carefully inspect lab data
coverage (see \code{plot_coverage} below). If lab data coverage is high (e.g.,
\>95\% of \code{genc_ids} have an entry in the lab table), individual \code{genc_ids}
may still not exist in the lab table (e.g., because lab testing was not
indicated) and individual lab columns might still have missing values.
Users should carefully consider how to handle these cases depending on the
context of their analyses.
}
\section{Warning!}{


\itemize{
\item{Data coverage checks should generally be performed on the whole
dataset, prior to applying any additional cohort inclusions/exclusions.}
\item{If you are an HPC4Health user and your datacut has been pre-filtered
based on certain inclusion/exclusion criteria (e.g., diagnosis codes),
please keep in mind that the coverage plots (see \code{plot_coverage}) may be
skewed in smaller/pre-filtered samples. Please reach out to the GEMINI
team if you need additional support.}
\item{Data coverage checks are particularly relevant for clinical data
tables (e.g., lab, pharmacy, radiology, transfusions, vitals, clinical
notes etc.).}
\item{This function should be used as a starting point for data coverage
checks, but users are advised to perform additional checks based on their
specific needs.}
}
}

\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
  dbname = "db",
  host = "domain_name.ca",
  port = 1234,
  user = getPass("Enter user:"),
  password = getPass("password")
)

cohort <- dbGetQuery(db, "SELECT genc_id FROM admdad;")

## run function with default flags to create all plots
# Note: This might take a while to run...
coverage <- data_coverage(dbcon, cohort, table = c("admdad", "radiology"))

# get flags per encounter based on encounter's discharge date
enc_flag <- coverage[["data"]][1] # coverage[["data"]]$coverage_flag_enc

# get data timeline (min-max dates)
data_timeline <- coverage[["data"]][2] # coverage[["data"]]$timeline_data

# get \% encounters with entry in each table by discharge month & hospital
prct_coverage <- coverage[["data"]][3] # coverage[["data"]]$coverage_data


## run function without any plots
# (will only return data.table with encounter-level flag)
coverage <- data_coverage(
  dbcon,
  cohort,
  table = c("admdad", "radiology"),
  plot_timeline = FALSE,
  plot_coverage = FALSE
)
}

}
