% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_covid_surge.R
\name{derive_covid_surge}
\alias{derive_covid_surge}
\title{Derive COVID Surge}
\usage{
derive_covid_surge(
  db,
  time_period = c("2015-04-01", "2023-06-30"),
  gim_only = FALSE,
  include_er = FALSE
)
}
\arguments{
\item{db}{(\code{DBIConnection})\cr
RPostgres DB connection.}

\item{time_period}{(\code{vecotr})\cr
A time period of interest to derive a surge index for.
For example, if the user would like to investigate the surge from
2020-10-01 to 2023-01-01, the user should supply
\code{c('2020-10-01', '2023-01-01')}.}

\item{gim_only}{(\code{logical})\cr
Flag denoting if the user would like the surge index for GIM patients only.
Note: This argument is set to FALSE by default.}

\item{include_er}{(\code{logical})\cr
Flag denoting if ER intervention codes are to be included in the calculation.
Note: This argument is set to FALSE by default.}
}
\value{
(\code{data.frame})\cr
\code{hospital_id} (\code{character}),\cr
\code{month_year} (\code{character}),\cr
\code{surge_index} (\code{numeric})
}
\description{
A function that derives the COVID-19 surge index for a given time period.
For time periods before 2020, or where COVID was not yet diagnosed, the surge index will be 0.
}
\examples{
\dontrun{
drv <- DBI::dbDriver("PostgreSQL")
db <- DBI::dbConnect(
    drv,
    dbname = "db_name",
    host = "domain_name.ca",
    port = 1234,
    user = getPass::getPass("Enter Username"),
    password = getPass::getPass("Enter Password")
)

covid_surge <- derive_covid_surge(db = db, 
time_period = c('2020-10-01','2023-01-01'),
gim_only = FALSE,
include_er = FALSE)

}

}
\references{
The formula to derive the COVID surge index is from :
https://doi.org/10.1001/jamanetworkopen.2023.23035

Data for hospital bed availability to be pulled from:
https://www.cihi.ca/sites/default/files/document/beds-staffed-and-in-operation-2018-2019-en-web.xlsx
}
