% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlson_comorbidity_index.R
\name{diagnoses_at_admission}
\alias{diagnoses_at_admission}
\title{Determine Diagnoses at Admission}
\usage{
diagnoses_at_admission(ipdiag, erdiag)
}
\arguments{
\item{ipdiag}{(\code{data.table})\cr
\code{ipdiagnosis} table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.}

\item{erdiag}{(\code{data.table})\cr
\code{erdiagnosis} table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.}
}
\value{

}
\description{
See CIHI's \href{https://www.cihi.ca/sites/default/files/document/hospital-standardized-mortality-ratio-meth-notes-en.pdf}{HSMR Methodology}.

Specifically page 19 for inclusion/exclusion criteria.
}
