% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_admdad}
\alias{dummy_admdad}
\title{Generated simulated administrative data}
\usage{
dummy_admdad(id, admtime)

dummy_admdad(id, admtime)
}
\arguments{
\item{id}{(\code{numeric})\cr
A single identifier that is repeated to match the length of \code{value}.}

\item{admtime}{(\code{character})\cr
In the format yyyy-mm-dd hh:mm. Corresponds to the admission time of the encounter.}
}
\value{
(\code{data.table})\cr
With the columns \code{id} and \code{admission_date_time} as described above.

(\code{data.table})\cr
With the columns \code{id} and \code{admission_date_time} as described above.
}
\description{
Designed to partially mimic the \code{admdad} table as defined in the
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.

Designed to partially mimic the \code{admdad} table as defined in the
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
}
\examples{
admdad <- dummy_admdad(1, "2023-01-02 00:00")

admdad <- dummy_admdad(1, "2023-01-02 00:00")

}
