% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_diag}
\alias{dummy_diag}
\title{Generate Simulated Diagnosis Data Table}
\usage{
dummy_diag(
  nid = 1000,
  n_hospitals = 10,
  cohort = NULL,
  ipdiagnosis = TRUE,
  diagnosis_type = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{nid}{(\code{integer})\cr Number of unique encounter IDs (\code{genc_id}) to simulate. Value must be greater than 0.
Optional when \code{cohort} is provided.}

\item{n_hospitals}{(\code{integer})\cr Number of hospitals to simulate in the resulting data table.
Optional when \code{cohort} is provided.}

\item{cohort}{(\code{data.frame or data.table})\cr Optional, the administrative data frame with the columns:
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{hospital_num} (\code{integer}): hospital ID
\code{cohort} takes precedence over parameters \code{nid} and\code{n_hospitals}.
When \code{cohort} is not NULL, \code{nid} and \code{n_hospitals} are ignored.
}}

\item{ipdiagnosis}{(\code{logical})\cr Default to "TRUE" and returns simulated "ipdiagnosis" table.
If FALSE, returns simulated "erdiagnosis" table.
See tables in \href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.}

\item{diagnosis_type}{(\verb{character vector})\cr The type(s) of diagnosis to return.
Possible diagnosis types are
("M", 1", "2", "3", "4", "5", "6", "9", "W", "X", and "Y"). Regardless of \code{diagnosis_type} input,
the \code{ipdiagnosis} table is defaulted to always return type "M" for the first row of each encounter.}

\item{seed}{(\code{integer})\cr Optional, a number to assign the seed to.}

\item{...}{Additional arguments for ICD code sampling scheme. See \code{sample_icd()} for details.}
}
\value{
(\code{data.table})\cr A data table containing simulated data of
\code{genc_id}, \verb{(er)_diagnosis_code}, \verb{(er)_diagnosis_type}, \code{hospital_num},
and other fields found in the respective diagnosis table.
}
\description{
This function generates simulated data table resembling \code{ipdiagnosis}
or \code{erdiagnosis} tables that can be used for testing or demonstration purposes.
It internally calls \code{sample_icd()} function to sample ICD-10 codes and
accepts arguments passed to \code{sample_icd()} for customizing the sampling scheme.
}
\details{
To ensure simulated table resembles "ip(er)diagnosis" table, the following characteristics are applied to fields:
\itemize{
\item \code{genc_id}: Numerical identification of encounters starting from 1.
The number of unique encounters is defined by \code{n}. The total number of rows is defined by \code{nrow},
where the number of rows for each encounter is random, but each encounter has at least one row.
\item \code{hospital_num}: Numerical identification of hospitals from 1 to 5.
All rows of an encounter are linked to a single hospital
\item \code{diagnosis_code}: "ipdiagnosis" table only. Simulated ICD-10 diagnosis codes.
Each encounter can be associated with multiple diagnosis codes in long format.
\item \code{diagnosis_type}: "ipdiagnosis" table only.
The first row of each encounter is consistently assigned to the diagnosis type "M".
For the remaining rows, if \code{diagnosis_type} is specified by users,
diagnosis types are sampled randomly from values provided;
if \code{diagnosis_type} is NULL, diagnosis types are sampled from
("1", "2", "3", "4", "5", "6", "9", "W", "X", and "Y"),
with sampling probability proportionate to their prevalence in the "ipdiagnosis" table.
\item \code{diagnosis_cluster}: "ipdiagnosis" table only.
Proportionally sampled from values that have a prevalence of more than 1\%
in the "diagnosis_cluster" field of the "ipdiagnosis" table, which are ("", "A", "B").
\item \code{diagnosis_prefix}: "ipdiagnosis" table only.
Proportionally sampled from values that have a prevalence of more than 1\%
in the "diagnosis_prefix" field of the "ipdiagnosis" table, which are ("", "N", "Q", "6").
\item \code{er_diagnosis_code}: "erdiagnosis" table only.
Simulated ICD-10 diagnosis codes.
Each encounter can be associated with multiple diagnosis codes in long format.
\item \code{er_diagnosis_type}: "erdiagnosis" table only.
Proportionally sampled from values that have a prevalence of more than 1\%
in the "er_diagnosis_type" field of the "erdiagnosis" table, which are ("", "M", "9", "3", "O").
}
}
\note{
The following fields \verb{(er)diagnosis_code}, \verb{(er)diagnosis_type}, \code{diagnosis_cluster}, \code{diagnosis_prefix}
are simulated independently.
Therefore, the simulated combinations may not reflect the interrelationships of these fields in actual data.
For example, specific diagnosis codes may be associated with specific diagnosis types,
diagnosis clusters, or diagnosis prefix in reality.
However, these relationships are not maintained for the purpose of generating dummy data.
Users require specific linkages between these fields should consider customizing
the output data or manually generating the desired combinations.
}
\examples{

### Simulate an erdiagnosis table for 5 unique subjects with total 20 records:
\dontrun{
set.seed(1)
erdiag <- dummy_diag(nid = 50, n_hospitals = 2, ipdiagnosis = F)
}

### Simulate an erdiagnosis table including data from `cohort`
cohort <- dummy_ipadmdad()
erdiag <- dummy_diag(cohort = cohort)

### Simulate an ipdiagnosis table with diagnosis codes starting with "E11":
\dontrun{
set.seed(1)
ipdiag <- dummy_diag(nid = 50, n_hospitals = 20, ipdiagnosis = T, pattern = "^E11")
}

### Simulate a ipdiagnosis table with random diagnosis codes in diagnosis type 3 or 6 only:
\dontrun{
set.seed(1)
ipdiag <- dummy_diag(nid = 50, n_hospitals = 10, diagnosis_type = (c("3", "6"))) \%>\%
  filter(diagnosis_type != "M") # remove default rows with diagnosis_type="M" from each ID
}

### Simulate a ipdiagnosis table with ICD-10-CA codes:
\dontrun{
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
  dbname = "db",
  host = "172.XX.XX.XXX",
  port = 1234,
  user = getPass("Enter user:"),
  password = getPass("password")
)

set.seed(1)
ipdiag <- dummy_diag(nid = 5, n_hospitals = 2, ipdiagnosis = T, dbcon = dbcon, source = "icd_lookup")
}

}
