% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_er}
\alias{dummy_er}
\title{Generate simulated ER data.}
\usage{
dummy_er(
  nid = 1000,
  n_hospitals = 10,
  time_period = c(2015, 2023),
  cohort = NULL,
  seed = NULL
)
}
\arguments{
\item{nid}{(\code{integer}) Number of unique encounter IDs to simulate. In this data table, each ID occurs once.
Optional when \code{cohort} is provided.}

\item{n_hospitals}{(\code{integer}) Number of hospitals in the simulated dataset. Optional when \code{cohort} is provided.}

\item{time_period}{(\code{vector})\cr A numeric or character vector containing the data range of the data
by years or specific dates in either format: ("yyyy-mm-dd", "yyyy-mm-dd") or (yyyy, yyyy).
The start date and end date will be (yyyy-01-01 and yyyy-12-31) if (yyyy, yyyy)
is the date range format provided. Optional when \code{cohort} is provided.}

\item{cohort}{(\verb{data.frame or data.table}): Optional, a data frame with the following columns:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID
\item \code{hospital_num} (\code{integer}): Mock hospital ID number
\item \code{admission_date_time} (\code{character}): The date and time of admission to the hospital with format "\%Y-\%m-\%d \%H:\%M"
\item \code{discharge_date_time} (\code{character}): The date and time of discharge from the hospital with format "\%Y-\%m-\%d \%H:\%M"
When \code{cohort} is not NULL, \code{nid}, \code{n_hospitals}, and \code{time_period} are ignored.
}}

\item{seed}{(\code{integer}) Optional, a number for setting the seed to get reproducible results.}
}
\value{
(\code{data.table}) A data.table object similar to the "er" table that contains the following fields:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID; integers starting from 1 or from \code{cohort} if provided
\item \code{hospital_num} (\code{integer}): Mock hospital ID; integers starting from 1 or from \code{cohort} if provided
\item \code{triage_date_time} (\code{character}): The date and time of triage with format "\%Y-\%m-\%d \%H:\%M"
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "er" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.

This function will return one triage date time for each encounter ID.
}
\examples{
cohort <- dummy_ipadmdad()
dummy_er(cohort = cohort, seed = 1)
dummy_er(nid = 10, n_hospitals = 1, seed = 2)
}
