% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/dummy_data.R
\docType{data}
\name{dummy_ipdiagnosis}
\alias{dummy_ipdiagnosis}
\title{Dummy ipdiagnosis data}
\description{
This mimics the GEMINI ipdiagnosis table. And is generated with the \code{dummy_diag} function.
See the \href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}
for details.
}
\keyword{internal}
