% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_lab}
\alias{dummy_lab}
\title{Generated simulated lab data}
\usage{
dummy_lab(id, omop, value, unit, mintime)
}
\arguments{
\item{id}{(\code{numeric})\cr
A single identifier that is repeated to match the length of \code{value}.}

\item{omop}{(\code{character})\cr
Codes corresponding to OMOP concept identifiers.}

\item{value}{(\code{numeric})\cr
Simulated result values for each lab test measurement.}

\item{unit}{(\code{character})\cr
Units corresponding to the particular lab test as defined by \code{omop}. It is repeated to match the length of \code{value}.}

\item{mintime}{(\code{character})\cr
In the format yyyy-mm-dd hh:mm. Earliest recorded test performed time.}
}
\value{
(\code{data.table})\cr
With the columns, \code{id}, \code{omop}, \code{value}, \code{unit}, and \code{collection_date_time} as described above.
}
\description{
Designed to mimic the most important elements of the GEMINI lab table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.
}
\examples{
lab <- dummy_lab(1, 3024641, c(7,8,15,30), "mmol/L", "2023-01-02 08:00")

}
