% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_locality}
\alias{dummy_locality}
\title{Generate simulated locality variables data}
\usage{
dummy_locality(
  dbcon = NULL,
  nid = 1000,
  n_hospitals = 10,
  cohort = NULL,
  da21uid = NULL,
  seed = NULL
)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A connection to the GEMINI database, used to access the 2021 Canadian census dissemination codes.
It is required when \code{da21uid} is missing. If \code{da21uid} is provided, then this input is ignored.}

\item{nid}{(\code{integer})\cr Number of unique encounter IDs to simulate. In this data table, each ID occurs once.
It is optional when \code{cohort} is provided.}

\item{n_hospitals}{(\code{integer})\cr Number of hospitals in simulated dataset.
It is optional when \code{cohort} is provided.}

\item{cohort}{(\verb{data.frame or data.table}) Optional, an existing data frame or data table similar to \code{admdad} in
GEMINI with at least the following columns:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID; integers starting from 1.
\item \code{hospital_num} (\code{integer}): Mock hospital ID; integers starting from 1.
If \code{cohort} is provided, \code{nid} and \code{n_hospital} inputs are not used.
}}

\item{da21uid}{(\code{integer} or \code{vector})\cr Allows the user to customize which dissemination area ID to include in
the output. It is required when \code{dbcon} is missing. It can be an integer or an integer vector.
If included, the \code{dbcon} is not used.}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results.}
}
\value{
(\code{data.table})\cr
A data.table object similar to the "locality_variables" table that contains the following fields:
\itemize{
\item \code{genc_id} (\code{integer}): Mock encounter ID; integers starting from 1.
\item \code{da21uid} (\code{integer}): Dissemination area ID based on 2021 Canadian census data using PCCF Version 8A
}
}
\description{
This function creates a synthetic dataset with a subset of variables that
are contained in the GEMINI "locality-variables" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.

Specifically, the function simulates dissemination area IDs (da21uid) based on Canadian census data for a
user-specified set of mock encounter and hospital IDs. To mimic GEMINI data characteristics, the majority
of simulated area IDs are drawn from Ontario and are clustered by hospital.
}
