% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_physicians}
\alias{dummy_physicians}
\title{Generate simulated physician data}
\usage{
dummy_physicians(nid = 1000, n_hospitals = 10, cohort = NULL, seed = NULL)
}
\arguments{
\item{n_hospitals}{(\code{integer})\cr Number of hospitals in simulated dataset.}

\item{cohort}{(\code{data.frame}) Optional, an existing data table similar to \code{admdad} in GEMINI
containing administrative information, at least the GEMINI encounter ID and hospital number.}

\item{seed}{(\code{integer})\cr Optional, a number to be used to set the seed for reproducible results.}

\item{n}{(\code{integer})\cr Number of unique encounter IDs to simulate.}
}
\value{
(\code{data.table})\cr A data.table object similar to the "physicians" table that contains the
following fields (if \code{cohort} is included, these are in addition to any columns in \code{cohort}):
\itemize{
\item \code{genc_id} (\code{integer}): GEMINI encounter ID
\item \code{hospital_num} (\code{integer}): Hospital ID number
\item \code{admitting_physician_gim} (\code{logical}): Whether the admitting physician attends a general medicine ward
\item \code{discharging_physician_gim} (\code{logical}): Whether the discharging physician attends a general medicine ward
\item \code{adm_phy_cpso_mapped} (\code{integer}): Unique hash of admitting physician CPSO Number
\item \code{mrp_cpso_mapped} (\code{integer}): Unique hash of most responsible physician (MRP) CPSO Number
\item \code{dis_phy_cpso_mapped} (\code{integer}): Unique hash of discharging physician CPSO Number
}
}
\description{
This function creates a dummy dataset with a subset of variables that
are contained in the GEMINI "physicians" table, as seen in
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}.
}
\examples{
dummy_physicians(nid = 1000, n_hospitals = 10, seed = 1)

}
