% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailty_score.R
\name{frailty_score}
\alias{frailty_score}
\title{Compute frailty score using ICD-10-CA}
\usage{
frailty_score(ipadmdad, ipdiagnosis, eddiagnosis)
}
\arguments{
\item{ipadmdad}{(\code{data.table})
\code{ipadmdad} table as defined in the \href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.
This table must contain the \code{genc_id} field}

\item{ipdiagnosis}{(\code{data.table})
\code{ipdiagnosis} table as defined in the \href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.
This table must contain the \code{genc_id} and \code{diagnosis_code} (as ICD-10-CA) fields in long format table only.
The diagnosis codes must be free from any punctuation or special characters.}

\item{eddiagnosis}{(\code{data.table})
\code{eddiagnosis} table as defined in the \href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.
This table must contain the \code{genc_id} and \code{er_diagnosis_code} (as ICD-10-CA) fields in long format table only.
The diagnosis codes must be free from any punctuation or special characters.}
}
\value{
\code{data.table}
This function returns a table with encounters identified by the \code{ipadmdad} table input and
additional derived numeric fields for \code{ip_frailty_score_derived}, \code{er_frailty_score_derived}
and \code{frailty_score_derived}
\itemize{
\item{ip_frailty_score_derived : }{Frailty risk score calculated only using IP diagnoses codes from \code{ipdiagnosis} table}
\item{er_frailty_score_derived : }{Frailty risk score calculated only using ED diagnoses codes from \code{eddiagnosis} table}
\item{frailty_score_derived : }{Frailty risk score calculated using both IP & ED diagnoses codes}
}
If frailty risk score is < 5, the patient is low-risk
If frailty risk score is between 5 and 15, the patient is intermediate risk
If frailty risk score is > 15.0, the patient is high risk
}
\description{
\code{frailty_score} returns frailty risk score for hospital admissions calculated based on ICD-10-CA diagnoses codes.
}
\details{
This function takes a list of ICD-10-CA and generates a numeric field corresponding to frailty risk
score for each admission.
This function can take any \code{data.table} with ICD-10-CA codes.
Historically, scores have been calculated based on partition of ICD-10-CA codes.

Researchers may want to focus on frailty risk scores based on in-patient (IP) diagnoses because
emergency department (ED) diagnoses have lower sensitivity/specificity for certain conditions and
are less reliably filled compared to IP diagnoses. In general, there will be overlap between IP and
ED diagnosis codes. ED diagnosis codes to compute frailty risk score can be used to evaluate frailty
risk upon admission to IP services.

This function does not differentiate between diagnosis types.
}
\note{
The original frailty risk score was developed based on ICD-10 codes instead of ICD-10-CA.
However, ICD-10-CA version codes are identical to the ICD-10 counterpart up to 4 character levels
and the frailty risk score only uses the first three characters of ICD-10.

Please refer to the references in this page for more details.
}
\section{Warning}{

This function returns a \code{data.table} with \code{genc_id} and three numeric fields.
The function will return NA values if \code{genc_id} in the \code{ipadmdad} table is not found
in the \code{ipdiagnosis} or \code{eddiagnosis} tables.

\code{frailty_score_derived} is NA if either \code{ip_frailty_score_derived} or \code{er_frailty_score_derived} are NA.
When one tries to left-join the output of this function with another table (another list of admissions in the left),
make sure list of admissions (or patient ids) aligns in both tables.
}

\references{
Gilbert T, Neuburger J, Kraindler J, et al. Development and validation of a
Hospital Frailty Risk Score focusing on older people in acute care settings using
electronic hospital records: an observational study. Lancet 2018; published online
April 26. http://dx.doi.org/10.1016/S0140-6736(18)30668-8.
}
