% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini_theme.R
\name{gemini_theme}
\alias{gemini_theme}
\title{GEMINI plot theme}
\usage{
gemini_theme(
  base_size = 14,
  base_family = "",
  show_grid = FALSE,
  aspect_ratio = 1
)
}
\arguments{
\item{bg_colour}{}
}
\description{
GEMINI plot theme
}
