% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gemini_theme}
\alias{gemini_theme}
\title{GEMINI plot theme}
\usage{
gemini_theme(
  base_size = 14,
  base_family = "",
  show_grid = FALSE,
  aspect_ratio = 1
)
}
\arguments{
\item{base_size}{(\code{numeric})\cr}

\item{base_family}{(\code{numeric})\cr}

\item{show_grid}{(\code{logical})\cr}

\item{aspect_ratio}{(\code{numeric})\cr
Aspect ratio of plot}
}
\value{
ggplot with specified theme
}
\description{
Common theme that can be applied to any ggplot object
}
\references{
https://emanuelaf.github.io/own-ggplot-theme.html
https://ggplot2.tidyverse.org/reference/theme.html
https://rpubs.com/mclaire19/ggplot2-custom-themes
https://themockup.blog/posts/2020-12-26-creating-and-using-custom-ggplot2-themes/
https://stackoverflow.com/questions/34522732/changing-fonts-in-ggplot2
http://www.cookbook-r.com/Graphs/Fonts/
https://github.com/wch/extrafont
https://github.com/koundy/ggplot_theme_Publication
}
