% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_fiscal_year.R
\name{hospital_fiscal_year}
\alias{hospital_fiscal_year}
\title{Get hospital fiscal year}
\usage{
hospital_fiscal_year(date_ymd)
}
\arguments{
\item{date_ymd}{a character vector of label of dates or date-times.
The date value must be in "yyyy-mm-dd" format}
}
\value{
vector with the same number of entries with corresponding fiscal years
to input dates.
Possible values of this field is "2010", "2011", "2012" and so on.
}
\description{
\code{hospital_fiscal_year} returns the corresponding hospital fiscal years for a set of dates provided.
}
\details{
Hospital fiscal year is defined as April ~ March. For example, in case of 2015-02-25, fiscal year will be 2014.
}
\section{Warning}{

NA values in returned vector indicates either missing date or
its format is incorrect, as all dates must be in yyyy-mm-dd format and precede the timestamp.
}

