% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_hospital_mortality.R
\name{in_hospital_mortality}
\alias{in_hospital_mortality}
\title{Compute in-hospital mortality using CIHI DAD}
\usage{
in_hospital_mortality(ipadmdad, suicide = FALSE)
}
\arguments{
\item{ipadmdad}{(\code{data.frame} or \code{data.table})\cr
Table with all relevant encounters of interest from DRM table "ipadmdad" (see
\href{https://geminimedicine.ca/the-gemini-database/}{GEMINI Data Repository Dictionary}).
Must contain two fields: \code{genc_id} and \code{discharge_disposition}.}

\item{suicide}{(\code{logical})\cr
a TRUE/FALSE argument indicating whether suicides in facility
should be counted towards in-hospital mortality rate or not.}
}
\value{
\code{data.table} with the same number of rows as input \code{ipadmdad}, with
additional derived boolean field labelled as \code{in_hospital_mortality_derived}.
Possible values are \code{TRUE}, \code{FALSE} or \code{NA}, where \code{NA} indicates missing
discharge disposition.
}
\description{
\code{in_hospital_mortality} returns whether a patient has deceased
in hospital based on the CIHI DAD field discharge disposition
}
\details{
This function takes the CIHI DAD "discharge disposition" field
(Group 05, Field 05 As of 19/20 CIHI DAD) and generates a boolean
variable indicating whether the patient has deceased during hospital stay.

The current version of the function captures the four possible
values indicating that a patient has deceased in hospital:
\itemize{
\item 7: Died (Deprecated post 2017)
\item 72: Died in Facility
\item 73: Medical Assistance in Dying
\item 74: Suicide in Facility (Excluded in calculation by default;
can be included by setting \code{suicide} argument to \code{TRUE})
}
}
\section{Warning}{

\code{NA} values in returned data.table indicate missing
discharge disposition field values.
}

