% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{length_of_stay}
\alias{length_of_stay}
\title{Compute hospital Length of Stay  (LoS)}
\usage{
length_of_stay(
  admdad,
  idvar1 = "genc_id",
  admvar1 = "admission_date_time",
  disvar1 = "discharge_date_time"
)
}
\arguments{
\item{admdad}{(\code{data.table})\cr
\code{admdad} table as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.

admdad is a data.table object equivalent of DRM table "admdad".
Table must contain three fields:
admission id, admission date and time and discharge date time.}

\item{idvar1}{a character string of label for primary key of "admdad" table, ideally as admission id}

\item{admvar1}{a character string of label equivalent to "admission date time".
Its value must be in "yyyy-mm-dd hh:mm" format}

\item{disvar1}{a character string of label equivalent to "discharge date time".
Its value must be in "yyyy-mm-dd hh:mm" format}
}
\value{
data.table with the same number of rows with input "admdad" with addtional
derived numeric fields labelled as "los_hrs_derived" and "los_days_derived".
}
\description{
\code{length_of_stay} returns Length of Stay for hospital admissions in unit of hours and days
}
\details{
Length of Stay is defined as the duration of hospital in-patient stay,
thus, calculated as (discharge date time - admission date time).
It is currently a major patient outcome in various research studies
and one of the physician performance indicators in MyPracticeReport.
Researchers might consider a few modifications to this metric.
One example will be adjusting for patient who receives palliative care.
If transfer to palliative unit within the hospital is trackable
(e.g. through room transfer data),
then it might be good idea to re-define discharge point to the palliative unit entrance.
This will reduce bias in metric since patient care under palliative order is very different.
}
\section{Warning}{

NA values in returned data.table indicates either missing admission/discharge date and time or
its format is incorrect
}

\section{Note}{

Function takes required fields labels as input, this was intended not to hardcode variable names
within the function. When field labels change in database, there is no need to go through the lines to change
the hardcoded variable names
}

