% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{length_of_stay}
\alias{length_of_stay}
\title{Compute hospital Length of Stay  (LoS)}
\usage{
length_of_stay(
  ipadmdad,
  adm_dtvar = "admission_date_time",
  dis_dtvar = "discharge_date_time"
)
}
\arguments{
\item{ipadmdad}{(\code{data.frame} or \code{data.table})\cr
Table with all relevant encounters of interest from DRM table "ipadmdad" (see
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}).
Must contain three fields: encounter ID (\code{genc_id}), and two date-time
variables corresponding to admission and discharge (typically
\code{admission_date_time} and \code{discharge_date_time}). Date-time variables must be
in "yyyy-mm-dd hh:mm" format.}

\item{adm_dtvar}{(\code{character})\cr
Character string defining the column name for admission date-time (typically
"admission_date_time").}

\item{dis_dtvar}{(\code{character})\cr
Character string defining the column name for discharge date-time (typically
"discharge_date_time").}
}
\value{
data.table with the same number of rows as input "ipadmdad", with additional
derived numeric fields labelled as "los_hrs_derived" and "los_days_derived".
}
\description{
\code{length_of_stay} returns Length of Stay for hospital admissions in unit of
hours and days.
}
\details{
Length of Stay is defined as the duration of hospital in-patient stay,
thus, calculated as (\code{discharge_date_time - admission_date_time}).
It is currently a major patient outcome in various research studies
and one of the physician performance indicators in MyPracticeReport.
Researchers might consider a few modifications to this metric.
One example will be adjusting for patient who receives palliative care.
If transfer to palliative unit within the hospital is trackable
(e.g. through room transfer data), then it might be good idea to re-define
discharge point to the palliative unit entrance. This will reduce bias in
metric since patient care under palliative order is very different.
}
\section{Warning}{

NA values in returned data.table indicates either missing admission/discharge
date and time or its format is incorrect.
}

\section{Note}{

The function takes two optional input arguments defining the admission and
discharge date-time variables of interest (by default \code{admission_date_time}
and \code{discharge_date_time}).
}

