% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_of_stay.R
\name{length_of_stay}
\alias{length_of_stay}
\title{Compute hospital Length of Stay  (LoS)}
\usage{
length_of_stay(
  admdad,
  admvar1 = "admission_date_time",
  disvar1 = "discharge_date_time"
)
}
\arguments{
\item{admdad}{(\code{data.table} or \code{data.frame})\cr
Table equivalent to DRM table "ipadmdad" as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.
Table must contain three fields: encounter ID (\code{genc_id}), and two date-time
variables corresponding to admission and discharge (typically
\code{admission_date_time} and \code{discharge_date_time}).}

\item{admvar1}{(\code{character})\cr
a character string equivalent to "admission date time".
Must be in "yyyy-mm-dd hh:mm" format.}

\item{disvar1}{(\code{character})\cr
a character string equivalent to "discharge date time".
Must be in "yyyy-mm-dd hh:mm" format.}
}
\value{
data.table with the same number of rows as input "admdad", with additional
derived numeric fields labelled as "los_hrs_derived" and "los_days_derived".
}
\description{
\code{length_of_stay} returns Length of Stay for hospital admissions in unit of
hours and days.
}
\details{
Length of Stay is defined as the duration of hospital in-patient stay,
thus, calculated as (discharge date time - admission date time).
It is currently a major patient outcome in various research studies
and one of the physician performance indicators in MyPracticeReport.
Researchers might consider a few modifications to this metric.
One example will be adjusting for patient who receives palliative care.
If transfer to palliative unit within the hospital is trackable
(e.g. through room transfer data), then it might be good idea to re-define
discharge point to the palliative unit entrance. This will reduce bias in
metric since patient care under palliative order is very different.
}
\section{Warning}{

NA values in returned data.table indicates either missing admission/discharge
date and time or its format is incorrect.
}

\section{Note}{

The function takes two optional input arguments defining the admission and
discharge date-time variables of interest (by default \code{admission_date_time}
and \code{discharge_date_time}).
}

