% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlaps.R
\name{loop_mlaps}
\alias{loop_mlaps}
\title{Loop mLAPS}
\usage{
loop_mlaps(
  db,
  cohort = NULL,
  hours_after_admission = 0,
  output_laps_components = FALSE
)
}
\arguments{
\item{db}{(\code{DBIConnection})\cr
RPostgres DB connection.}

\item{cohort}{(\code{numeric})\cr
A numeric vector of \code{genc_ids} to restrict the calculation to.}

\item{hours_after_admission}{(\code{numeric})\cr
Consider lab tests collected \strong{up to} \code{hours_after_admission} hours after admission in the
calculation.}

\item{output_laps_components}{(\code{logical})\cr
Does not aggregate the score and instead outputs for each LAPS component (test), its contribution
to the overall score.}
}
\value{
(\code{data.frame})\cr
If \code{output_laps_components == TRUE}:
\code{genc_id} (\code{numeric}),\cr
\code{test_type_mapped_omop} (\code{character}),\cr
\code{mlaps} (\code{numeric}) max score for this test.

If \code{output_laps_components == FALSE}:
\code{genc_id} (\code{numeric}),\cr
\code{mlaps} (\code{numeric}) sum of max scores for each relevant test for this encounter.
}
\description{
A wrapper around the \code{mlaps()} function which breaks down the calculation of LAPS on a
large number of encounters by hospital-year. This avoids memory issues that can be caused
by loading large chunks of the lab table.
}
