% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlaps.R
\name{mlaps}
\alias{mlaps}
\title{mLAPS}
\usage{
mlaps(admdad, lab, hours_offset = 0, componentwise = FALSE)
}
\arguments{
\item{admdad}{(\code{data.frame})\cr
Table equivalent to a subset of the \code{admdad} table defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.}

\item{lab}{(\code{data.table}, \code{data.frame})\cr
Table equivalent to a subset of the \code{lab} table defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.}

\item{hours_offset}{(\code{numeric})\cr
Consider lab tests collected \strong{up to} \code{hours_after_admission} hours after admission in the}

\item{componentwise}{(\code{logical})\cr
Does not aggregate the score and instead outputs for each LAPS component (test), its contribution
to the overall score.}
}
\value{
(\code{data.frame})\cr
If \code{componentwise == TRUE}:
\code{genc_id} (\code{numeric}),\cr
\code{test_type_mapped_omop} (\code{character}),\cr
\code{mlaps} (\code{numeric}) max score for this test.

If \code{componentwise == FALSE}:
\code{genc_id} (\code{numeric}),\cr
\code{mlaps} (\code{numeric}) sum of max scores for each relevant test for this encounter.
}
\description{
mLAPS
}
\details{
Laboratory based Acute Physiology Score (LAPS) uses 14 lab test values.
In this modified version, High senstive Troponin tests are ignored and treated as normal.
}
\examples{

}
\references{
https://pubmed.ncbi.nlm.nih.gov/18388836/
}
