% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_imaging.r
\name{n_imaging}
\alias{n_imaging}
\title{Compute the number of radiology tests per encounter}
\usage{
n_imaging(dbcon, cohort, exclude_ed = FALSE)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database. \code{DBI} connection is recommended
as \code{odbc} connection may cause connection issues in certain environment.}

\item{cohort}{(\code{data.frame} or \code{data.table})
Cohort table with all relevant encounters of interest, where each row
corresponds to a single encounter. Must contain GEMINI Encounter ID
(\code{genc_id}).}

\item{exclude_ed}{(\code{logical})
Whether to exclude tests in emergency department. When set to \code{TRUE}, only
tests being performed in inpatient settings are counted. When set to \code{FALSE},
tests in both ED and in-patient settings will be counted. To distinguish
tests in ED and in-patient settings, following
the methodology implemented in
\href{https://www.hqontario.ca/Portals/0/documents/qi/practice-reports/general-medicine-sample-report.html#imaging-qi}{My Practice Report},
\code{ordered_date_time} is compared against \code{admission_date_time} to identify
imaging tests in ED. When \code{ordered_date_time} is not available,
\code{performed_date_time} is used instead. If both \code{ordered_date_time} and
\code{performed_date_time} are not available, the test will not be counted in the
output.}
}
\value{
data.table with the same number of rows as input "cohort", with additional
derived numeric fields labelled as "n_img_xray_derived",  "n_img_ct_derived",
"n_img_mri_derived",  "n_img_us_derived", "n_img_other_derived",
"n_img_int_derived" and "n_img_ct_mri_us_derived".
}
\description{
\code{n_imaging} returns the number of radiology tests for hospital admission.
}
\details{
This function takes a list of admissions and a GEMINI databse connection to
generate numeric fields counting the number of different radiology tests for
each admission.

Imaging table in the database should include field that classifies each test
into below 7 categories.

\itemize{
\item{1 : }{X-Ray}
\item{2 : }{CT}
\item{3 : }{MRI}
\item{4 : }{Ultrasound}
\item{5 : }{Other}
\item{6 : }{Interventional radiology procedures}
\item{7 : }{Echo (Excluded)}
}

Therefore, this function should be run after some standardization efforts on
imaging table (performed by GEMINI team). Currently, the number of advanced
radiology tests is one of the performance metrics in MyPracticeReport. CT,
MRI and Ultrasound are considered advance imaging tests, which can be
retrieved by the derived variable (\code{n_img_ct_mri_us_derived}).
}
\note{
Currently, the function does not take data availability into account. For
patients without imaging tests, the function will return 0 in result columns.
User should check radiology data availability and decide whether the imputed
\code{0}s are appropriate or not.
}
\section{Warning}{

Function returns data.table with id field and several numeric fields. By
design, function will not return any NA values. When one tries to left-join
the output of this function with another table (another list of admissions in
the left), make sure list of admissions (or patient) aligns in both tables.
}

\section{Note}{

The function takes optional inputs specifying the field label for the
variables corresponding to date-time of the performed imaging test and
mapped imaging modality, respectively. This is to avoid hardcoding variable
names within the function in case field labels differ between databases.
}

