% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_imaging.r
\name{n_imaging}
\alias{n_imaging}
\title{Compute the number of radiology tests per encounter}
\usage{
n_imaging(
  ipadmdad,
  imaging,
  dtvar = "performed_date_time",
  mapvar = "modality_mapped"
)
}
\arguments{
\item{ipadmdad}{(\code{data.table} or \code{data.frame})\cr
Table equivalent to DRM table "ipadmdad". Must contain \code{genc_id}.}

\item{imaging}{(\code{data.table} or \code{data.frame})\cr
Table equivalent to DRM table "radiology" as defined in the
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary}.
Table must contain three fields: \code{genc_id}, imaging performed date time (see
\code{dtvar}, in "yyyy-mm-dd hh:mm" format), and a field that standardizes the
imaging test type (see \code{mapvar}).}

\item{dtvar}{(\code{character})
Name of the column in \code{imaging} table containing date time of performed
imaging test (usually \code{"performed_date_time"}).}

\item{mapvar}{(\code{character})
Name of the column in \code{imaging} table containing the imaging test type
(usually \code{"modality_mapped"}).}
}
\value{
data.table with the same number of rows as input "ipadmdad", with additional
derived numeric fields labelled as "n_img_xray_derived",  "n_img_ct_derived",
"n_img_mri_derived",  "n_img_us_derived", "n_img_other_derived",
"n_img_int_derived" and "n_img_ct_mri_us_derived".
}
\description{
\code{n_imaging} returns the number of radiology tests for hospital admission.
}
\details{
This function takes a list of admissions and DRM table equivalent of imaging
table to generate numeric fields counting the number of different radiology
tests for each admission.

Imaging table in the database should include field that classifies each test
into below 7 categories.

\itemize{
\item{1 : }{X-Ray}
\item{2 : }{CT}
\item{3 : }{MRI}
\item{4 : }{Ultrasound}
\item{5 : }{Other}
\item{6 : }{Interventional radiology procedures}
\item{7 : }{Echo (Excluded)}
}

Therefore, this function should be run after some standardization efforts on
imaging table (performed by GEMINI team). Currently, the number of advanced
radiology tests is one of the performance metrics in MyPracticeReport. CT,
MRI and Ultrasound are considered advance imaging tests, which can be
retrieved by the derived variable (\code{n_img_ct_mri_us_derived}).

Any tests being performed in both ED and in-patient settings will be counted.
}
\section{Warning}{

Function returns data.table with id field and several numeric fields. By
design, function will not return any NA values. When one tries to left-join
the output of this function with another table (another list of admissions in
the left), make sure list of admissions (or patient) aligns in both tables.
}

\section{Note}{

The function takes optional inputs specifying the field label for the
variables corresponding to date time of the performed imaging test and
mapped imaging modality, respectively. This is to avoid hardcoding variable
names within the function in case field labels differ between databases.
}

