% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_routine_bloodwork.r
\name{n_routine_bloodwork}
\alias{n_routine_bloodwork}
\title{Compute the number of routine bloodwork tests per encounter}
\usage{
n_routine_bloodwork(con, ipadmdad, lab_table = "lab")
}
\arguments{
\item{con}{(\code{DBIConnection})\cr
A database connection to any GEMINI Database}

\item{ipadmdad}{(\code{data.table} or \code{data.frame})\cr
Table equivalent to DRM table "ipadmdad". Must contain \code{genc_id}.}

\item{lab_table}{(\code{character})\cr
Name of table corresponding to DRM lab table (typically "lab" or
"lab_subset")}
}
\value{
data.table object with the same number of rows as input "ipadmdad", with
additional derived numeric field labelled as "n_routine_bloodwork_derived"
}
\description{
\code{n_routine_bloodwork} returns the number of routine bloodwork (Sodium and
Hemoglobin tests) for each hospital admission. Sodium and Heamoglobin tests
are defined by OMOP codes. Sodium is the code 3019550. Hemoglobin is 3000963.
}
\details{
This function takes a list of admissions and DRM table equivalent of lab
table (see
\href{https://drive.google.com/uc?export=download&id=1iwrTz1YVz4GBPtaaS9tJtU0E9Bx1QSM5}{GEMINI Data Repository Dictionary})
to generate numeric fields counting the number of Sodium and Hemoglobin tests
for each admission.

Lab table in the database should include field that classifies each test into
a few different categories. Therefore, this function should be run after some
standardization efforts on lab table (performed by GEMINI). Currently, the
number of routine bloodwork tests is one of the performance metrics in
MyPracticeReport.

Function automatically removes tests without valid numeric result value.
Exceptions are boundary result value (e.g. ">120" mmol/L), and these
are still counted as valid tests.

Any tests being performed in both ED and in-patient settings will be counted.
}
\section{Warning}{

Function returns data.table with id field \code{genc_id} and one numeric field
indicating the number of bloodwork tests per encounter. By design,
function will not return any NA values. If a \code{genc_id} does not have any
entries in the "lab" table, the admission gets assigned 0 number of tests.
When one tries to left-join the output of this function with another table
(another list of admissions in the left), make sure list of admissions (or
patient) aligns in both tables

This function requires mappings by a Subject Matter Expert to ensure that all
tests are mapped to
3019550 - Sodium (Moles/volume) in Serum or Plasma, and
3000963 - Hemoglobin (Mass/volume) in Blood
}

