% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_ses.R
\name{neighborhood_ses}
\alias{neighborhood_ses}
\title{Obtain commonly used neighbourhood-level socioeconomic status (SES) variables}
\usage{
neighborhood_ses(dbcon, cohort, census_year)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{cohort}{(\code{data.table} | \code{data.frame})\cr
Table with all relevant encounters of interest, where each row corresponds to
a single encounter. Must contain GEMINI Encounter ID (\code{genc_id}).}

\item{census_year}{(\code{numeric} | \code{character})\cr
Statistics Canada census year. Only 2016 or 2021 are valid inputs.}
}
\value{
(\code{data.frame} | \code{data.table})\cr
This function returns a data.table where each row corresponds to a `genc_id`` from the user-provided cohort input, together with the following columns:
\itemize{
\item DA (dissemination area) the encounter resides in: da21uid (da16uid) for 2021 (2016) census year
\item Neighbourhood-level income: qnatippe, qnbtippe, qaatippe, qabtippe, atippe, btippe. While qnatippe and qnbtippe are quintiles calculated based on national income distribution, qaatippe and qabtippe are quintiles constructed separately for each census metropolitan area (CMA), census agglomeration (CA) or residual area within each province. Atippe, btippe are numeric.
\item \href{https://www12.statcan.gc.ca/census-recensement/2016/ref/dict/pop127-eng.cfm}{Visible minority}: vismin_pct
\item \href{https://www12.statcan.gc.ca/census-recensement/2021/dp-pd/prof/details/page.cfm?LANG=E&GENDERlist=1,2,3&STATISTIClist=4&HEADERlist=23&SearchText=Canada&DGUIDlist=2021A000011124}{Immigration status}: immsta_pct
\item \href{https://www150.statcan.gc.ca/n1/pub/81-004-x/2010001/def/posteducation-educpost-eng.htm}{Post-secondary education (>15)}: ed_15over_postsec_pct
\item Post-secondary education (25-64): ed_25to64_postsec_pct
\item Ontario Marginalization Index (numeric): (instability_da16, deprivation_da16, dependency_da16, ethniccon_da16) or (households_dwellings_DA21, material_resources_DA21, age_labourforce_DA21, racialized_NC_pop_DA21)
\item Ontario Marginalization Index (quintile): (instability_q_da16, deprivation_q_da16, dependency_q_da16, ethniccon_q_da16) or (households_dwellings_q_DA21, material_resources_q_DA21, age_labourforce_q_DA21, racialized_NC_pop_q_DA21)
}
}
\description{
The \code{neighborhood_ses()} function derives neighborhood-level SES variables for a given encounter based on the dissemination area they reside in. All variables returned by this function are based on Statistics Canada Census data and the Ontario Marginalization Index.

Note that the function output differs based on the census year (2016 or 2021) provided by the user.
}
\section{Statistics Canada Census}{

All SES variables are sourced from the Statistics Canada Census. The census is collected every 5 years and provides a detailed statistical portrait of communities across Canada, including information about income, education, ethnicity, and immigrant status.

The following census variables are based on a long-form questionnaire, which is only administered to 25\% of households:
\itemize{
\item Education
\item Visible minority status
}

In 2016, immigrant status was also based on the long-form questionnaire but since 2021 has been sourced from Immigration, Refugees and Citizenship Canada.

In both 2016 & 2021, information about household income was sourced from the Canadian Revenue Agency.

All census data are collected by dissemination area (DA), which typically covers a population of 400-700 people. To enable linkage between GEMINI data and DA-level information, the DA of a given encounter was derived from their postal code using the \href{https://www150.statcan.gc.ca/n1/en/catalogue/92-154-X}{Postal Code Conversion File Plus (PCCF+)} program.
}

\section{Ontario Marginalization Index (On-Marg)}{

On-Marg is a neighborhood-level index measuring marginalization differences between areas based on a subset of 42 variables from the Statistics Canada census.

The index was derived from a principal component factor analysis, which resulted in 18 indicators along the following 4 dimensions:
\enumerate{
\item Households and dwellings: Measures housing density and characteristics of family structure (e.g., living alone, \% dwellings not owned)
\item Material resources: Measures access to basic material needs (e.g., employment rate)
\item Age and labour force: Includes indicators such as the \% of seniors (65+), children, and those that are not part of the labour force
\item Racialized and newcomer populations: Measures the \% of people who are newcomers or identify as a visible minority
}

In the 2016 version of ON-Marg, the 4 dimensions were called "Residential instability", "Material deprivation", "Dependency", and "Ethnic concentration" respectively. The dimensions were renamed in 2021 to avoid deficit-based language and better reflect the census measures associated with each dimension.

All ON-Marg variables are available as continuous factor scores as well as quintiles, where higher values represent a higher degree of marginalization.
}

\examples{
\dontrun{
drv <- dbDriver("PostgreSQL")
dbcon <- DBI::dbConnect(drv,
  dbname = "db",
  host = "domain_name.ca",
  port = 1234,
  user = getPass("Enter user:"),
  password = getPass("password")
)

neighborhood_ses_table <- neighborhood_ses(dbcon, cohort, 2021)
}

}
\references{
\href{https://www12.statcan.gc.ca/census-recensement/2021/ref/98-304/2021001/chap1-eng.cfm}{Statistics Canada Census.}
\href{https://www.publichealthontario.ca/-/media/Event-Presentations/2023/09/ontario-marginalization-index-updates-products.pdf?rev=07baae2569164c17abaa18464075aa20&sc_lang=en}{ON-Marg guide 2021}
\href{https://www.publichealthontario.ca/-/media/documents/U/2018/userguide-on-marg.pdf}{ON-Marg guide 2016}
\href{https://www.cihi.ca/sites/default/files/document/toolkit-area-level-measurement-pccf-en.pdf}{Measuring Health Inequalities: A Toolkit Area-Level Equity Stratifiers Using PCCF and PCCF+}
}
