% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_color_palettes}
\alias{plot_color_palettes}
\title{Plot GEMINI color palettes}
\usage{
plot_color_palettes(plot_palettes = "all")
}
\arguments{
\item{plot_palettes}{(\code{character} or \code{numeric})\cr
Which palettes to plot. By default, plots all palettes.}
}
\description{
Plots all available GEMINI color palettes.
}
\seealso{
\code{vignette("plotting_theme", package = "Rgemini")}
}
