% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot descriptive summary statistics of multiple variables}
\usage{
plot_summary(data, plot_vars = NULL, show_stats = TRUE, color = "lightblue")
}
\arguments{
\item{data}{(\code{data.frame} | \code{data.table})\cr
Table containing data to be plotted.}

\item{plot_vars}{(\code{character} | \code{list})\cr
Character vector or list of variables to be plotted.}

\item{show_stats}{(\code{logical})\cr
Flag indicating whether to show descriptive stats above each plot.}

\item{color}{(\code{character})\cr
Plotting color. Default is R's built-in \code{"lightblue"}.}
}
\value{
(\code{ggplot})\cr A ggplot figure containing histograms of all variables
specified in \code{vars}.
}
\description{
This function plots distributions (histograms/bar plots) and basic summary
statistics (e.g., median \link{Q1, Q3}, \% missing etc.) for multiple variables.
}
\section{Note}{

These plots are not meant as publication-ready figures. Instead, the goal of
this function is to provide a quick and easy means to visually inspect the
data and obtain information about distributional properties of a wide range
of variables, requiring just a single line of code.
}

\examples{

# simulate GEMINI data table
admdad <- dummy_data(n = 10000,
                         n_hospitals = 20,
                         time_period = c(2015, 2022)
                         )

plot_summary(data = admdad,
             plot_vars = c("age", "gender", "discharge_disposition", "number_of_alc_days")




}
