% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{Plot theme for ggplots}
\usage{
plot_theme(base_size = 12, base_family = "sans", show_grid = NULL, ...)
}
\arguments{
\item{base_size}{(\code{numeric})\cr
Numeric input specifying the base font size, which will be passed to
\code{ggthemes::theme_foundation()}.}

\item{base_family}{(\code{character})\cr
Base font family (e.g., "sans", "mono", "serif"). Run \code{windowsFonts()} to
check for available options. Input will be passed to
\code{ggthemes::theme_foundation()}.}

\item{show_grid}{(\code{character})\cr
Character inputs specifying whether to show "major" or "minor" grid lines.
Default is \code{NULL} to not show any grid lines. Both "major" and "minor" grid
lines can be plotted by providing a character vector, i.e.,
\code{show_grid = c("major", "minor")}.}

\item{...}{\cr
Additional arguments passed to \code{ggplot2::theme()}.}
}
\value{
ggplot with specified theme
}
\description{
Common plotting theme that can be applied to any ggplot object. The theme is
based on \code{ggthemes::theme_foundation()}, with some additional features.
}
\seealso{
\code{vignette("plotting_theme", package = "Rgemini")}
}
