% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_suppression.R
\name{render_cell_suppression.continuous}
\alias{render_cell_suppression.continuous}
\title{Render Cell Suppression (Continuous)}
\usage{
render_cell_suppression.continuous(x, ...)
}
\arguments{
\item{x}{(\code{numeric})\cr
A continuous variable to summarize.}

\item{...}{\cr
Further arguments, passed to \code{table1:::stats.apply.rounding()}.}

\item{continuous_fn}{(\code{character})\cr
A character string specifying the summary statistics to display.
Accepts "mean", "median", or c("mean", "median") to display both. Defaults to "mean".}
}
\value{
named (\code{character})\cr
Concatenated with \code{""} to shift values down one row for proper alignment.
}
\description{
This is a custom renderer for \code{table1} continuous variables which performs GEMINI
"cell suppression" for any variable levels which contain fewer than 6 observations.
}
\examples{
x <- 1:6
render_cell_suppression.continuous(x)

y <- 1:2
render_cell_suppression.continuous(y)

Use in `table1`:
dat <- expand.grid(id=1:10, treat=c("Treated", "Placebo"))
dat$age <- runif(nrow(dat), 10, 50)
label(dat$age) <- "Age"

table1(~ age | treat, data=dat,
       render.continuous = render_cell_suppression.continuous,
       continuous_fn = c("mean", "median"), # to display mean and median simultaneously
       digits=2)


}
