% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxnorm_query.R
\name{rxnorm_query}
\alias{rxnorm_query}
\title{Retrieve rows from GEMINI pharmacy data matching specified drug term(s).}
\usage{
rxnorm_query(
  dbcon,
  class_input = NA,
  drug_input = NA,
  cohort = NULL,
  return_unmatched = FALSE,
  detailed_search = NULL,
  return_drug_list = NULL
)
}
\arguments{
\item{dbcon}{PostgreSQL connection class
The connection object for the desired database and user. Obtained with odbc::dbConnect()
Supported database versions: drm_cleandb_v3_1_0 / H4H_v5 or newer
Older DBs lack row_num so it will not work}

\item{class_input}{string or list of strings
Classification keyword or specific ids like ATC are accepted.}

\item{drug_input}{string or list of strings
Generic or brand name is accepted. Spelling is normalized.}

\item{cohort}{(optional, default: 'all') data frame of the cohort.}

\item{return_unmatched}{(mandatory, default: FALSE)
If TRUE, the function will output a list with two vectors:
(1) First vector called matched_rows: All the matched entries in a long format same output as if the argument return_unmatched is FALSE.
(2) Second vector called unmatched_rows: all unmatched pharmacy rows (wide format with genc_id, 6 identifying columns, and row_num) based on a list of genc_ids inputted.
If FALSE, the function will by default output matched entries in a long format for every genc_id wiht the following columns: genc_id, search_type, raw_input, rxnorm_match}

\item{detailed_search}{(optional, default: TRUE) logical
If TRUE, search for every related concept to each selected drug, instead of just the selected drugs.
This will greatly expand the search but will sometimes match related concepts that are not desired.}

\item{return_drug_list}{(optional, default: FALSE) logical
Outputs the search drug list instead of searching}
}
\value{
For unmatched_row = FALSE

The GEMINI pharmacy dataframe for matched rows, or NA if cancelled.
An additional 'rxnorm_match' column is added specifying what rxnorm matched the row to.
}
\description{
Retrieve rows from GEMINI pharmacy data matching specified drug term(s).
}
\examples{
\dontrun{
diabetes_orders_condensed <- rxnorm_query(dbcon = con,
                                       drug_input = c("metformin", "insulin"),
                                       return_unmatched = FALSE)
}

}
