% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{Get Season}
\usage{
season(date)
}
\arguments{
\item{date}{(\code{Date})\cr}
}
\value{
(\code{character})\cr
One of "Spring", "Summer", "Fall", "Winter".
}
\description{
Given a date, return the named season.
}
\examples{
season(as.Date(Sys.Date()))

}
