% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_table}
\alias{temp_table}
\title{Write temp tables}
\usage{
temp_table(dbcon, data, table_name = "temp_table", analyze = TRUE)
}
\arguments{
\item{dbcon}{(\code{DBIConnection})\cr
A database connection to any GEMINI database.}

\item{data}{(\code{data.table} or \code{data.frame})\cr
Data table to be written to DB as temp table.}

\item{table_name}{(\code{data.table} or \code{data.frame})\cr
Name of temporary table in DB (default = "temp_table").}

\item{analyze}{(\code{logical})\cr
Whether or not to use SQL Analyze statement to further improve
query efficiency (recommended).}
}
\description{
This function writes temporary tables to the database to improve query
efficiency.
Temporary tables are automatically removed once the user disconnects from the
database.
}
\examples{
\dontrun{
temp_table(dbcon, data.table(genc_id = c(1, 2, 3)))
}
}
